Generic parameter class

The GenericParam class allows to store generic parameter information such as the name of the function that was/has to be called (slot fun) and its arguments (slot args). This object is used to track the process history of the data processings of an XCMSnExp object. This is in contrast to e.g. the CentWaveParam object that is passed to the actual processing method.

Usage

GenericParam(fun = character(), args = list())

Arguments

fun

character representing the name of the function.

args

list (ideally named) with the arguments to the function.

Value

The GenericParam function returns a GenericParam object.

Slots

fun

character specifying the function name.

args

list (ideally named) with the arguments to the function.

See also

Author

Johannes Rainer

Examples

prm <- GenericParam(fun = "mean")

prm <- GenericParam(fun = "mean", args = list(na.rm = TRUE))