Store `XcmsExperiment` object as .RData file

The `RDataParam` class and method allow users to save an `XcmsExperiment` object as an .RData file with a chosen filename. The object gets exported using [`save()`] function. This `param` class and method are part of the possible dispatch of the generic function `storeResults`.

Usage

RDataParam(fileName = tempfile())

# S4 method for class 'XcmsExperiment,RDataParam'
storeResults(object, param)

Arguments

fileName

for `RDataParam` `character(1)`, defining the file name. The default will be `tempfile()`.

object

MsExperiment or XcmsExperiment The data object that needs to be saved.

param

The parameter object selecting and configuring the format for saving. It can be one of the following classes: RDataParam, PlainTextParam, or MzTabMParam.

Value

for `RDataParam`: a `RDataParam` class. `storeResults` does not return anything but saves the object to a RData file.

See also

Author

Philippine Louail

Examples

## Load a test data set with detected peaks
faahko_sub <- loadXcmsData("faahko_sub2")

## Define param 
param <- RDataParam(fileName = "example_xcms_object")

## Save as RData
storeResults(object = faahko_sub, param = param)