absent-methods {xcms}	R Documentation

Determine which peaks are absent / present in a sample class

Description

Determine which peaks are absent / present in a sample class

Arguments

object
xcmsSet-class object

class
Name of a sample class from sampclass

minfrac
minimum fraction of samples necessary in the class to be absent/present

Details

Determine which peaks are absent / present in a sample class The functions treat peaks that are only present because of fillPeaks correctly, i.e. does not count them as present.

Value

An logical vector with the same length as nrow(groups(object)).

Methods

object = "xcmsSet"

absent(object, ...) present(object, ...)

See Also

group diffreport

[Package xcms version 3.0.2 Index]