chromatographic-peak-detection {xcms}	R Documentation

Chromatographic peak detection methods.

Description

The findChromPeaks methods perform the chromatographic peak detection on LC/GC-MS data and are part of the modernized xcms user interface.

The implemented peak detection methods in chromatographic space are:

centWave

chromatographic peak detection using the centWave method. See centWave for more details.

centWave with predicted isotopes

peak detection using a two-step centWave-based approach considering also feature isotopes. See centWaveWithPredIsoROIs for more details.

matchedFilter

peak detection in chromatographic space. See matchedFilter for more details.

massifquant

peak detection using the Kalman filter-based method. See massifquant for more details.

MSW

single-spectrum non-chromatography MS data peak detection. See MSW for more details.

Author(s)

Johannes Rainer

See Also

findPeaks for the old peak detection methods.

plotChromPeaks to plot identified chromatographic peaks for one file.

highlightChromPeaks to highlight identified chromatographic peaks in an extracted ion chromatogram plot.

Other peak detection methods: findChromPeaks-centWaveWithPredIsoROIs, findChromPeaks-centWave, findChromPeaks-massifquant, findChromPeaks-matchedFilter, findPeaks-MSW

[Package xcms version 3.8.1 Index]