dirname {xcms}	R Documentation

Change the file path of an OnDiskMSnExp object

Description

dirname allows to get and set the path to the directory containing the source files of the OnDiskMSnExp (or XCMSnExp) object.

Usage

## S4 method for signature 'OnDiskMSnExp'
dirname(path)

## S4 replacement method for signature 'OnDiskMSnExp'
dirname(path) <- value

Arguments

path
OnDiskMSnExp.

value
character of length 1 or length equal to the number of files defining the new path to the files.

Author(s)

Johannes Rainer

[Package xcms version 3.8.2 Index]