Filter intensities of a Chromatogram object

Filter a Chromatogram() object removing data points with intensities below a user provided threshold. If intensity is a numeric value, the returned chromatogram will only contain data points with intensities > intensity. In addition it is possible to provide a function to perform the filtering. This function is expected to take the input Chromatogram (object) and to return a logical vector with the same length then there are data points in object with TRUE for data points that should be kept and FALSE for data points that should be removed. See examples below.

# S4 method for Chromatogram
filterIntensity(object, intensity = 0, ...)

# S4 method for MChromatograms
filterIntensity(object, intensity = 0, ...)

Arguments

object
Chromatogram() object.

intensity
numeric(1) or function to use to filter the intensities. See description for details.

...
optional additional parameters. These will be passed to the filter function if such a function is specified with parameter intensity.

Value

filtered Chromatogram object.

Author

Johannes Rainer

Examples

chr1 <- Chromatogram(rtime = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
    intensity = c(3, 5, 14, 30, 24, 6, 2, 1, 1, 0))

## Remove data points with intensities below 10
res <- filterIntensity(chr1, 10)
intensity(res)

#> [1] 14 30 24

## Remove data points with an intensity lower than 10% of the maximum
## intensity in the Chromatogram
filt_fun <- function(x, prop = 0.1) {
    x@intensity >= max(x@intensity, na.rm = TRUE) * prop
}
res <- filterIntensity(chr1, filt_fun)
intensity(res)

#> [1]  3  5 14 30 24  6

## Remove data points with an intensity lower than half of the maximum
res <- filterIntensity(chr1, filt_fun, prop = 0.5)
intensity(res)

#> [1] 30 24