filtfft {xcms}	R Documentation

Apply an convolution filter using an FFT

Description

Expands a vector to the length of the filter and then convolutes it using two successive FFTs.

Usage

filtfft(y, filt)

Arguments

y
numeric vector of data to be filtered

filt
filter with length nextn(length(y))

Value

A numeric vector the same length as y.

Author(s)

Colin A. Smith, csmith@scripps.edu

[Package xcms version 3.0.2 Index]