findChromPeaks,Chromatogram,MatchedFilterParam-method {xcms}	R Documentation

matchedFilter-based peak detection in purely chromatographic data

Description

findChromPeaks on a Chromatogram or Chromatograms object with a MatchedFilterParam parameter object performs matchedFilter-based peak detection on purely chromatographic data. See matchedFilter for details on the method and MatchedFilterParam for details on the parameter class. Note that not all settings from the MatchedFilterParam will be used. See peaksWithMatchedFilter() for the arguments used for peak detection on purely chromatographic data.

Usage

## S4 method for signature 'Chromatogram,MatchedFilterParam'
findChromPeaks(object, param,
  ...)

Arguments

object
a Chromatogram or Chromatograms object.

param
a MatchedFilterParam object specifying the settings for the peak detection. See peaksWithMatchedFilter() for the description of arguments used for peak detection.

...
currently ignored.

Value

If called on a Chromatogram object, the method returns a matrix with the identified peaks. See peaksWithMatchedFilter() for details on the matrix content.

Author(s)

Johannes Rainer

See Also

peaksWithMatchedFilter() for the downstream function and matchedFilter for details on the method.

Examples

od <- readMSData(system.file("cdf/KO/ko15.CDF", package = "faahKO"),
    mode = "onDisk")

## Extract chromatographic data for a small m/z range
chr <- chromatogram(od, mz = c(272.1, 272.3))[1, 1]

## Identify peaks with default settings
xchr <- findChromPeaks(chr, MatchedFilterParam())

## Plot the identified peaks
plot(xchr)

## Modify the settings
mfp <- MatchedFilterParam(fwhm = 60)
xchr <- findChromPeaks(chr, mfp)

plot(xchr)

[Package xcms version 3.6.1 Index]