findneutral {xcms}	R Documentation

Find neutral losses in xcmsFragment objects

Description

This is a method to find a neutral loss with a ppm window in a xcmsFragment object

Usage

findneutral(object, find, ppmE=25, print=TRUE)

Arguments

object
xcmsFragment object type

find
The neutral loss to be found

ppmE
the ppm error window for searching

print
If we should print a nice little report

Details

The method searches for a given neutral loss in an xcmsFragment object type given a certain ppm error window. The neutral losses are generated between neighbouring ions. The resulting data frame shows the whole scan in which the neutral loss was found.

Value

A data frame with the following columns:

PrecursorMz
The precursor m/z of the neutral losses

MSnParentPeakID
An index ID of the location of the precursor peak in the xcmsFragment object

msLevel
The level of the found fragment ion

rt
the Retention time of the found ion

mz
the actual m/z of the found fragment ion

intensity
The intensity of the fragment ion

sample
Which sample the fragment ion came from

GroupPeakMSn
an ID if the peaks were grouped by an xcmsSet grouping

CollisionEnergy
The collision energy of the precursor scan

Author(s)

H. Paul Benton, hpbenton@scripps.edu

References

H. Paul Benton, D.M. Wong, S.A.Strauger, G. Siuzdak "XCMS^2" Analytical Chemistry 2008

See Also

findMZ,

Examples

	## Not run: 
	library(msdata)
	mzdatapath <- system.file("iontrap", package = "msdata")
	mzdatafiles<-list.files(mzdatapath, pattern = "extracted.mzData", recursive = TRUE, full.names = TRUE)
	xs <- xcmsSet(mzdatafiles, method = "MS1")
	##takes only one file from the file set
	xfrag <- xcmsFragments(xs)
	found<-findneutral(xfrag, 58.1455, 50)
	
## End(Not run)

[Package xcms version 3.0.2 Index]