getSpec-methods {xcms}	R Documentation

Get average m/z and intensity values for multiple mass scans

Description

Return full-resolution averaged data from multiple mass scans.

Arguments

object
the xcmsRaw object

...
arguments passed to profRange used to sepecify the spectral segments of interest for averaging

Details

Based on the mass points from the spectra selected, a master unique list of masses is generated. Every spectra is interpolated at those masses and then averaged.

Value

A matrix with two columns:

mz
m/z values

intensity
intensity values

Methods

object = "xcmsRaw"

getSpec(object, ...)

See Also

xcmsRaw-class, profRange, getScan

[Package xcms version 3.6.1 Index]