Peak grouping based on time dimension peak densities

This method performs performs correspondence (chromatographic peak grouping) based on the density (distribution) of identified peaks along the retention time axis within slices of overlapping mz ranges. All peaks (from the same or from different samples) being close on the retention time axis are grouped into a feature (peak group).

The PeakDensityParam class allows to specify all settings for the peak grouping based on peak densities along the time dimension. Instances should be created with the PeakDensityParam() constructor.

sampleGroups,sampleGroups<-: getter and setter for the sampleGroups slot of the object. Its length should match the number of samples in the experiment and it should not contain NAs.

bw,bw<-: getter and setter for the bw slot of the object.

minFraction,minFraction<-: getter and setter for the minFraction slot of the object.

minSamples,minSamples<-: getter and setter for the minSamples slot of the object.

binSize,binSize<-: getter and setter for the binSize slot of the object.

maxFeatures,maxFeatures<-: getter and setter for the maxFeatures slot of the object.

groupChromPeaks,XCMSnExp,PeakDensityParam: performs correspondence (peak grouping within and across samples) within in mz dimension overlapping slices of MS data based on the density distribution of the identified chromatographic peaks in the slice along the time axis.

The correspondence analysis can be performed on chromatographic peaks of any MS level (if present and if chromatographic peak detection has been performed for that MS level) defining features combining these peaks. The MS level can be selected with the parameter msLevel. By default, calling groupChromPeaks will remove any previous correspondence results. This can be disabled with add = TRUE, which will add newly defined features to already present feature definitions.

PeakDensityParam(
  sampleGroups = numeric(),
  bw = 30,
  minFraction = 0.5,
  minSamples = 1,
  binSize = 0.25,
  maxFeatures = 50
)

# S4 method for PeakDensityParam
show(object)

# S4 method for PeakDensityParam
sampleGroups(object)

# S4 method for PeakDensityParam
sampleGroups(object) <- value

# S4 method for PeakDensityParam
bw(object)

# S4 method for PeakDensityParam
bw(object) <- value

# S4 method for PeakDensityParam
minFraction(object)

# S4 method for PeakDensityParam
minFraction(object) <- value

# S4 method for PeakDensityParam
minSamples(object)

# S4 method for PeakDensityParam
minSamples(object) <- value

# S4 method for PeakDensityParam
binSize(object)

# S4 method for PeakDensityParam
binSize(object) <- value

# S4 method for PeakDensityParam
maxFeatures(object)

# S4 method for PeakDensityParam
maxFeatures(object) <- value

# S4 method for XCMSnExp,PeakDensityParam
groupChromPeaks(object, param, msLevel = 1L, add = FALSE)

Arguments

sampleGroups
A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

bw
numeric(1) defining the bandwidth (standard deviation ot the smoothing kernel) to be used. This argument is passed to the [density() method.

minFraction
numeric(1) defining the minimum fraction of samples in at least one sample group in which the peaks have to be present to be considered as a peak group (feature).

minSamples
numeric(1) with the minimum number of samples in at least one sample group in which the peaks have to be detected to be considered a peak group (feature).

binSize
numeric(1) defining the size of the overlapping slices in mz dimension.

maxFeatures
numeric(1) with the maximum number of peak groups to be identified in a single mz slice.

object
For groupChromPeaks: an XCMSnExp object containing the results from a previous peak detection analysis (see findChromPeaks()).

For all other methods: a `PeakDensityParam` object.

value
The value for the slot.

param
A PeakDensityParam object containing all settings for the peak grouping algorithm.

msLevel
integer(1) (default msLevel = 1L) defining the MS level on which the correspondence should be performed. It is required that chromatographic peaks of the respective MS level are present.

add
logical(1) (default add = FALSE) allowing to perform an additional round of correspondence (e.g. on a different MS level) and add features to the already present feature definitions.

Value

The PeakDensityParam function returns a PeakDensityParam class instance with all of the settings specified for chromatographic peak alignment based on peak densities. Note that argument sampleGroups is mandatory and should represent either the sample grouping in the experiment. It's length has to match the number of sample in the experiments.

For groupChromPeaks: a XCMSnExp object with the results of the correspondence analysis. The definition of the resulting mz-rt features can be accessed with the featureDefinitions() method

Slots

.__classVersion__,sampleGroups,bw,minFraction,minSamples,binSize,maxFeatures

See corresponding parameter above. .__classVersion__ stores the version from the class. Slots values should exclusively be accessed via the corresponding getter and setter methods listed above.

Note

These methods and classes are part of the updated and modernized xcms user interface. All of the settings to the algorithm can be passed with a PeakDensityParam object.

References

Colin A. Smith, Elizabeth J. Want, Grace O'Maille, Ruben Abagyan and Gary Siuzdak. "XCMS: Processing Mass Spectrometry Data for Metabolite Profiling Using Nonlinear Peak Alignment, Matching, and Identification" Anal. Chem. 2006, 78:779-787.

See also

Examples

## Create a PeakDensityParam object
p <- PeakDensityParam(binSize = 0.05, sampleGroups = c(1, 1, 2, 2))
## Change hte minSamples slot
minSamples(p) <- 3
p

#> Object of class:  PeakDensityParam 
#> Parameters:
#>  sampleGroups: numeric of length 4 
#>  bw: 30 
#>  minFraction: 0.5 
#>  minSamples: 3 
#>  binSize: 0.05 
#>  maxFeatures: 50 

##############################
## Chromatographic peak detection and grouping.
##
## Below we perform first a peak detection (using the matchedFilter
## method) on some of the test files from the faahKO package followed by
## a peak grouping using the density method.
library(faahKO)
library(MSnbase)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)

## Reading 2 of the KO samples
raw_data <- readMSData(fls[1:2], mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

## Perform the chromatographic peak detection using the matchedFilter method.
mfp <- MatchedFilterParam(snthresh = 20, binSize = 1)
res <- findChromPeaks(raw_data, param = mfp)

head(chromPeaks(res))

#>             mz mzmin mzmax       rt    rtmin    rtmax      into      intf  maxo
#> CP001 205.0000   205 205.0 2784.635 2770.550 2800.284 1778568.9 3589494.1 84280
#> CP002 205.0000   205 205.0 2784.635 2770.550 2800.284 1778568.9 3589494.1 84280
#> CP003 207.0821   207 207.1 2712.647 2698.562 2726.731  380873.0  725545.6 18800
#> CP004 207.0821   207 207.1 2712.647 2698.562 2726.731  380873.0  725657.0 18800
#> CP005 233.0390   233 233.1 3019.378 3005.293 3033.462  399145.3  746572.1 19752
#> CP006 233.0390   233 233.1 3019.378 3005.293 3033.462  399145.3  746587.0 19752
#>            maxf i       sn sample
#> CP001 194326.10 1 88.11283      1
#> CP002 194326.10 1 87.30418      1
#> CP003  39851.35 1 25.05928      1
#> CP004  39851.56 1 26.10968      1
#> CP005  41455.94 1 28.58299      1
#> CP006  41456.03 1 30.40953      1

## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

#> 
#>   1   2 
#> 251 299 

## Performing the chromatographic peak grouping. Assigning all samples to
## the same sample group.
fdp <- PeakDensityParam(sampleGroups = rep(1, length(fileNames(res))))
res <- groupChromPeaks(res, fdp)

#> Processing 3195 mz slices ... 

#> OK

## The definition of the features (peak groups):
featureDefinitions(res)

#> DataFrame with 243 rows and 10 columns
#>           mzmed     mzmin     mzmax     rtmed     rtmin     rtmax    npeaks
#>       <numeric> <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> FT001   200.100   200.100   200.100   2876.97   2876.97   2876.97         1
#> FT002   201.091   201.091   201.091   3141.44   3141.44   3141.44         2
#> FT003   205.000   205.000   205.000   2787.76   2784.64   2790.89         4
#> FT004   207.090   207.082   207.098   2715.78   2712.65   2718.91         4
#> FT005   233.036   233.033   233.039   3021.73   3019.38   3024.07         4
#> ...         ...       ...       ...       ...       ...       ...       ...
#> FT239   593.500   593.500   593.500   3715.78   3715.78   3715.78         1
#> FT240   594.313   594.313   594.313   3603.11   3603.11   3603.11         2
#> FT241   596.334   596.317   596.334   3815.94   3797.16   3815.94         3
#> FT242   597.371   597.371   597.371   3815.94   3815.94   3815.94         1
#> FT243   599.313   599.313   599.313   3743.95   3743.95   3743.95         2
#>              X1         peakidx  ms_level
#>       <numeric>          <list> <integer>
#> FT001         1             252         1
#> FT002         1         253,254         1
#> FT003         2   1,  2,255,...         1
#> FT004         2   3,  4,257,...         1
#> FT005         2   5,  6,259,...         1
#> ...         ...             ...       ...
#> FT239         1             547         1
#> FT240         1         247,248         1
#> FT241         2     249,548,549         1
#> FT242         1             550         1
#> FT243         1         250,251         1

## Using the featureValues method to extract a matrix with the
## intensities of the features per sample.
head(featureValues(res, value = "into"))

#>        ko15.CDF  ko16.CDF
#> FT001        NA  451600.7
#> FT002        NA  272872.0
#> FT003 1778568.9 1567038.1
#> FT004  380873.0  449392.6
#> FT005  399145.3  356951.3
#> FT006  252282.0        NA

## The process history:
processHistory(res)

#> [[1]]
#> Object of class "XProcessHistory"
#>  type: Peak detection 
#>  date: Thu Apr 30 09:42:18 2020 
#>  info:  
#>  fileIndex: 1,2 
#>  Parameter class: MatchedFilterParam 
#>  MS level(s) 1 
#> 
#> [[2]]
#> Object of class "XProcessHistory"
#>  type: Peak grouping 
#>  date: Thu Apr 30 09:42:20 2020 
#>  info:  
#>  fileIndex: 1,2 
#>  Parameter class: PeakDensityParam 
#>  MS level(s) 1 
#>