High resolution peak grouping for single spectra samples

This method performs high resolution correspondence for single spectra samples.

The MzClustParam class allows to specify all settings for the peak grouping based on the mzClust algorithm. Instances should be created with the MzClustParam constructor.

sampleGroups,sampleGroups<-: getter and setter for the sampleGroups slot of the object.

ppm,ppm<-: getter and setter for the ppm slot of the object.

absMz,absMz<-: getter and setter for the absMz slot of the object.

minFraction,minFraction<-: getter and setter for the minFraction slot of the object.

minSamples,minSamples<-: getter and setter for the minSamples slot of the object.

groupChromPeaks,XCMSnExp,MzClustParam: performs high resolution peak grouping for single spectrum metabolomics data.

MzClustParam(
  sampleGroups = numeric(),
  ppm = 20,
  absMz = 0,
  minFraction = 0.5,
  minSamples = 1
)

# S4 method for MzClustParam
sampleGroups(object)

# S4 method for MzClustParam
sampleGroups(object) <- value

# S4 method for MzClustParam
ppm(object)

# S4 method for MzClustParam
ppm(object) <- value

# S4 method for MzClustParam
absMz(object)

# S4 method for MzClustParam
absMz(object) <- value

# S4 method for MzClustParam
minFraction(object)

# S4 method for MzClustParam
minFraction(object) <- value

# S4 method for MzClustParam
minSamples(object)

# S4 method for MzClustParam
minSamples(object) <- value

# S4 method for XCMSnExp,MzClustParam
groupChromPeaks(object, param, msLevel = 1L)

Arguments

sampleGroups

A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

ppm

numeric(1) representing the relative mz error for the clustering/grouping (in parts per million).

absMz

numeric(1) representing the absolute mz error for the clustering.

minFraction

numeric(1) defining the minimum fraction of samples in at least one sample group in which the peaks have to be present to be considered as a peak group (feature).

minSamples

numeric(1) with the minimum number of samples in at least one sample group in which the peaks have to be detected to be considered a peak group (feature).

object

For groupChromPeaks: an XCMSnExp object containing the results from a previous chromatographic peak detection analysis (see findChromPeaks()).

For all other methods: a `MzClustParam` object.

value

The value for the slot.

param

A MzClustParam object containing all settings for the peak grouping algorithm.

msLevel

integer(1) defining the MS level. Currently only MS level 1 is supported.

Value

The MzClustParam function returns a MzClustParam class instance with all of the settings specified for high resolution single spectra peak alignment.

For groupChromPeaks: a XCMSnExp object with the results of the peak grouping step (i.e. the features). These can be accessed with the featureDefinitions() method.

Slots

sampleGroups,ppm,absMz,minFraction,minSamples

See corresponding parameter above.

Note

These methods and classes are part of the updated and modernized xcms user interface which will eventually replace the group() methods. All of the settings to the algorithm can be passed with a MzClustParam object.

Calling groupChromPeaks on an XCMSnExp object will cause all eventually present previous correspondence results to be dropped.

References

Saira A. Kazmi, Samiran Ghosh, Dong-Guk Shin, Dennis W. Hill and David F. Grant
Alignment of high resolution mass spectra: development of a heuristic approach for metabolomics.
Metabolomics, Vol. 2, No. 2, 75-83 (2006)

See also

Examples

## Loading a small subset of direct injection, single spectrum files
library(msdata)
fticrf <- list.files(system.file("fticr-mzML", package = "msdata"),
                    recursive = TRUE, full.names = TRUE)
fticr <- readMSData(fticrf[1:2], msLevel. = 1, mode = "onDisk")

## Disable parallel processing for this example
register(SerialParam())

## Perform the MSW peak detection on these:
p <- MSWParam(scales = c(1, 7), peakThr = 80000, ampTh = 0.005,
             SNR.method = "data.mean", winSize.noise = 500)
fticr <- findChromPeaks(fticr, param = p)

head(chromPeaks(fticr))
#>            mz    mzmin    mzmax rt rtmin rtmax    into     maxo        sn intf
#> CP01 403.2367 403.2279 403.2447 -1    -1    -1 4735258 372259.4 22.975342   NA
#> CP02 409.1845 409.1755 409.1936 -1    -1    -1 4158404 310572.1 20.597978   NA
#> CP03 413.2677 413.2585 413.2769 -1    -1    -1 6099006 435462.6 27.217229   NA
#> CP04 414.2719 414.2652 414.2786 -1    -1    -1 1916658 157613.1  8.777433   NA
#> CP05 423.2363 423.2266 423.2459 -1    -1    -1 2708391 174252.7 14.745275   NA
#> CP06 427.2681 427.2583 427.2779 -1    -1    -1 6302089 461385.6 32.468892   NA
#>           maxf sample
#> CP01  814693.1      1
#> CP02  731557.2      1
#> CP03 1018994.8      1
#> CP04  280826.9      1
#> CP05  435858.5      1
#> CP06 1124549.4      1

## Now create the MzClustParam parameter object: we're assuming here that
## both samples are from the same sample group.
p <- MzClustParam(sampleGroups = c(1, 1))

fticr <- groupChromPeaks(fticr, param = p)
#> 8.00  16.00  24.00  28.00  32.00  40.00  48.00  52.00  56.00  60.00  64.00  72.00  80.00  84.00  88.00  96.00  100.00  

## Get the definition of the features.
featureDefinitions(fticr)
#> DataFrame with 17 rows and 9 columns
#>          mzmed     mzmin     mzmax     rtmed     rtmin     rtmax    npeaks
#>      <numeric> <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> FT01   403.237   403.237   403.237        -1        -1        -1         2
#> FT02   409.184   409.184   409.185        -1        -1        -1         2
#> FT03   413.267   413.267   413.268        -1        -1        -1         2
#> FT04   414.272   414.272   414.272        -1        -1        -1         1
#> FT05   419.213   419.213   419.213        -1        -1        -1         1
#> ...        ...       ...       ...       ...       ...       ...       ...
#> FT13   438.240   438.240   438.240        -1        -1        -1         2
#> FT14   443.241   443.241   443.241        -1        -1        -1         1
#> FT15   443.336   443.336   443.336        -1        -1        -1         1
#> FT16   445.293   445.292   445.293        -1        -1        -1         2
#> FT17   447.297   447.297   447.297        -1        -1        -1         1
#>             X1 peakidx
#>      <numeric>  <list>
#> FT01         2    1,13
#> FT02         2    14,2
#> FT03         2    15,3
#> FT04         1       4
#> FT05         1      16
#> ...        ...     ...
#> FT13         2   10,22
#> FT14         1      11
#> FT15         1      23
#> FT16         2   24,12
#> FT17         1      25