groupnames-methods {xcms}	R Documentation

Generate unque names for peak groups

Description

Allow linking of peak group data between classes using unique group names that remain the same as long as no re-grouping occurs.

Arguments

object
the xcmsSet or xcmsEIC object

mzdec
number of decimal places to use for m/z

rtdec
number of decimal places to use for retention time

template
a character vector with existing group names whose format should be emulated

Value

A character vector with unique names for each peak group in the object. The format is M[m/z]T[time in seconds].

Methods

object = "xcmsSet"

(object, mzdec = 0, rtdec = 0, template = NULL)

object = "xcmsEIC"

(object)

See Also

xcmsSet-class, xcmsEIC-class

[Package xcms version 3.0.2 Index]