isolationWindowTargetMz,OnDiskMSnExp-method {xcms}	R Documentation

Extract isolation window target m/z definition

Description

isolationWindowTargetMz extracts the isolation window target m/z definition for each spectrum in object.

Usage

## S4 method for signature 'OnDiskMSnExp'
isolationWindowTargetMz(object)

Arguments

object
OnDiskMSnExp object.

Value

a numeric of length equal to the number of spectra in object with the isolation window target m/z or NA if not specified/available.

Author(s)

Johannes Rainer

[Package xcms version 3.8.2 Index]