levelplot-methods {xcms}	R Documentation

Plot log intensity image of a xcmsRaw object

Description

Create an image of the raw (profile) data m/z against retention time, with the intensity color coded.

Arguments

x
xcmsRaw object.

log
Whether the intensity should be log transformed.

col.regions
The color ramp that should be used for encoding of the intensity.

rt
wheter the original (rt="raw") or the corrected (rt="corrected") retention times should be used.

...
Arguments for profRange.

Methods

x = "xcmsRaw"

levelplot(x, log=TRUE, col.regions=colorRampPalette(brewer.pal(9, "YlOrRd"))(256), ...)

x = "xcmsSet"

levelplot(x, log=TRUE, col.regions=colorRampPalette(brewer.pal(9, "YlOrRd"))(256), rt="raw", ...)

Author(s)

Johannes Rainer, johannes.rainer@eurac.edu

See Also

xcmsRaw-class, xcmsSet-class

[Package xcms version 3.8.2 Index]