medianFilter {xcms}	R Documentation

Apply a median filter to a matrix

Description

For each element in a matix, replace it with the median of the values around it.

Usage

medianFilter(x, mrad, nrad)

Arguments

x
numeric matrix to median filter

mrad
number of rows on either side of the value to use for median calculation

nrad
number of rows on either side of the value to use for median calculation

Value

A matrix whose values have been median filtered

Author(s)

Colin A. Smith, csmith@scripps.edu

Examples

mat <- matrix(1:25, nrow=5)
mat
medianFilter(mat, 1, 1)

[Package xcms version 3.8.1 Index]