plotScan-methods {xcms}	R Documentation

Plot a single mass scan

Description

Plot a single mass scan using the impulse representation. Most useful for centroided data.

Arguments

object
the xcmsRaw object

scan
integer with number of scan to plot

mzrange
numeric vector of length >= 2 whose range will be used to select masses to plot

ident
logical, use mouse to interactively identify and label individual masses

Methods

object = "xcmsRaw"

plotScan(object, scan, mzrange = numeric(), ident = FALSE)

See Also

xcmsRaw-class

[Package xcms version 3.8.2 Index]