plotSpec-methods {xcms}	R Documentation

Plot mass spectra from the profile matrix

Description

Uses the pre-generated profile mode matrix to plot mass spectra over a specified retention time range.

Arguments

object
the xcmsRaw object

ident
logical, use mouse to identify and label peaks

vline
numeric vector with locations of vertical lines

...
arguments passed to profRange

Value

If ident == TRUE, an integer vector with the indecies of the points that were identified. Otherwise a two-column matrix with the plotted points.

Methods

object = "xcmsRaw"

plotSpec(object, ident = FALSE, vline = numeric(0), ...)

See Also

xcmsRaw-class

[Package xcms version 3.0.2 Index]