plotTIC-methods {xcms}	R Documentation

Plot total ion count

Description

Plot chromatogram of total ion count. Optionally allow identification of target peaks and viewing/identification of individual spectra.

Arguments

object
the xcmsRaw object

ident
logical, use mouse to identify and label chromatographic peaks

msident
logical, use mouse to identify and label spectral peaks

Value

If ident == TRUE, an integer vector with the indecies of the points that were identified. Otherwise a two-column matrix with the plotted points.

Methods

object = "xcmsRaw"

plotTIC(object, ident = FALSE, msident = FALSE)

See Also

xcmsRaw-class

[Package xcms version 3.6.1 Index]