plotrt-methods {xcms}	R Documentation

Plot retention time deviation profiles

Description

Use corrected retention times for each sample to calculate retention time deviation profiles and plot each on the same graph.

Arguments

object
the xcmsSet object

col
vector of colors for plotting each sample

ty
vector of line and point types for plotting each sample

leg
logical plot legend with sample labels

densplit
logical, also plot peak overall peak density

Methods

object = "xcmsSet"

plotrt(object, col = NULL, ty = NULL, leg = TRUE, densplit = FALSE)

See Also

xcmsSet-class, retcor

[Package xcms version 3.0.2 Index]