profMethod-methods {xcms}	R Documentation

Get and set method for generating profile data

Description

These methods get and set the method for generating profile (matrix) data from raw mass spectral data. It can currently be bin, binlin, binlinbase, or intlin.

Methods

object = "xcmsRaw"

profMethod(object)

See Also

xcmsRaw-class, profMethod, profBin, plotSpec, plotChrom, findPeaks

[Package xcms version 3.0.2 Index]