rawEIC-methods {xcms}	R Documentation

Get extracted ion chromatograms for specified m/z range

Description

Generate extracted ion chromatogram for m/z values of interest. The raw data is used in contrast to getEIC which uses data from the profile matrix (i.e. values binned along the M/Z dimension).

Arguments

object
xcmsRaw object

mzrange
m/z range for EIC

rtrange
retention time range for EIC

scanrange
scan range for EIC

Value

A list of :

scan
scan number

intensity
added intensity values

Methods

object = "xcmsRaw"

rawEIC(object, mzrange = numeric(), rtrange = numeric(), scanrange = numeric())

Author(s)

Ralf Tautenhahn

See Also

xcmsRaw-class

[Package xcms version 3.8.2 Index]