retexp {xcms}	R Documentation

Set retention time window to a specified width

Description

Expands (or contracts) the retention time window in each row of a matrix as defined by the retmin and retmax columns.

Usage

retexp(peakrange, width = 200)

Arguments

peakrange
maxtrix with columns retmin and retmax

width
new width for the window

Value

The altered matrix.

Author(s)

Colin A. Smith, csmith@scripps.edu

See Also

getEIC

[Package xcms version 3.6.1 Index]