specDist.peakCount-methods {xcms}	R Documentation

a Distance function based on matching peaks

Description

This method calculates the distance of two sets of peaks by just returning the number of matching peaks (m/z-values).

Usage

specDist.peakCount(peakTable1, peakTable2, mzabs=0.001, mzppm=10, symmetric=FALSE)

Arguments

peakTable1
a Matrix containing at least m/z-values, row must be called "mz"

peakTable2
the matrix for the other mz-values

mzabs
maximum absolute deviation for two matching peaks

mzppm
relative deviations in ppm for two matching peaks

symmetric
use symmetric pairwise m/z-matches only, or each match

Methods

peakTable1 = "matrix", peakTable2 = "matrix"

specDist.peakCount(peakTable1, peakTable2, mzppm=10,symmetric=FALSE )

Author(s)

Joachim Kutzera, jkutzer@ipb-halle.de

[Package xcms version 3.8.2 Index]