split.xcmsRaw {xcms}	R Documentation

Divide an xcmsRaw object

Description

Divides the scans from a xcmsRaw object into a list of multiple objects. MS$^n$ data is discarded.

Arguments

x
xcmsRaw object

f
factor such that factor(f) defines the scans which go into the new xcmsRaw objects

drop
logical indicating if levels that do not occur should be dropped (if 'f' is a 'factor' or a list).

...
further potential arguments passed to methods.

Value

A list of xcmsRaw objects.

Methods

xr = "xcmsRaw"

split(x, f, drop = TRUE, ...)

Author(s)

Steffen Neumann, sneumann(at)ipb-halle.de

See Also

xcmsRaw-class

[Package xcms version 3.6.1 Index]