useOriginalCode {xcms}	R Documentation

Enable usage of old xcms code

Description

This function allows to enable the usage of old, partially deprecated code from xcms by setting a corresponding global option. See details for functions affected.

Usage

useOriginalCode(x)

Arguments

x
logical(1) to specify whether or not original old code should be used in corresponding functions. If not provided the function simply returns the value of the global option.

Details

The functions/methods that will be affected by this are:

do_findChromPeaks_matchedFilter

Value

logical(1) indicating whether old code is being used.

Note

Usage of old code is strongly dicouraged. This function is thought to be used mainly in the transition phase from xcms to xcms version 3.

Author(s)

Johannes Rainer

[Package xcms version 3.0.2 Index]