writeMSData,XCMSnExp,character-method {xcms}	R Documentation

Export MS data to mzML/mzXML files

Description

writeMSData exports mass spectrometry data in mzML or mzXML format. If adjusted retention times are present, these are used as retention time of the exported spectra.

Usage

## S4 method for signature 'XCMSnExp,character'
writeMSData(object, file,
  outformat = c("mzml", "mzxml"), copy = FALSE,
  software_processing = NULL, ...)

Arguments

object
XCMSnExp object with the mass spectrometry data.

file
character with the file name(s). The length of this parameter has to match the number of files/samples of object.

outformat
character(1) defining the format of the output files ( either "mzml" or "mzxml").

copy
logical(1) if metadata (data processing, software used, original file names etc) should be copied from the original files.

software_processing
optionally provide specific data processing steps. See documentation of the software_processing parameter of mzR::writeMSData().

...
Additional parameters to pass down to the writeMSData() function in the MSnbase package, such as outformat to specify the output format ("mzml" or "mzxml") or copy to specify whether general information from the original MS data files (such as data processing, software etc) should be copied to the new files.

Author(s)

Johannes Rainer

See Also

writeMSData() function in the MSnbase package.

[Package xcms version 3.8.1 Index]