xcms-deprecated {xcms}	R Documentation

Deprecated functions in package ‘xcms’

Description

These functions are provided for compatibility with older versions of ‘xcms’ only, and will be defunct at the next release.

Details

The following functions/methods are deprecated.

xcmsPapply: this function is no longer available and the use of bplapply is suggested.

profBin, profBinM, profBinLin, profBinLinM, profBinLinBase, profBinLinBaseM have been deprecated and binYonX in combination with imputeLinInterpol should be used instead.

extractChromatograms: replaced by chromatogram.

plotChromatogram: replaced by plot method for Chromatogram or Chromatograms objects.

[Package xcms version 3.0.2 Index]