xcmsPeaks-class {xcms}	R Documentation

A matrix of peaks

Description

A matrix of peak information. The actual columns depend on how it is generated (i.e. the findPeaks method).

Objects from the Class

Objects can be created by calls of the form new("xcmsPeaks", ...).

Slots

.Data:

The matrix holding the peak information

Extends

Class "matrix", from data part. Class "array", by class "matrix", distance 2. Class "structure", by class "matrix", distance 3. Class "vector", by class "matrix", distance 4, with explicit coerce.

Methods

None yet. Some utilities for working with peak data would be nice.

Author(s)

Michael Lawrence

See Also

findPeaks for detecting peaks in an xcmsRaw.

[Package xcms version 3.8.2 Index]