xcmsSource-methods {xcms}	R Documentation

Create an xcmsSource object in a flexible way

Description

Users can define alternate means of reading data for xcmsRaw objects by creating new implementations of this method.

Methods

signature(object = "xcmsSource")

Pass the object through unmodified.

Author(s)

Daniel Hackney, dan@haxney.org

See Also

xcmsSource

[Package xcms version 3.8.2 Index]