GenericParam-class {xcms}	R Documentation

Generic parameter class

Description

The GenericParam class allows to store generic parameter information such as the name of the function that was/has to be called (slot fun) and its arguments (slot args). This object is used to track the process history of the data processings of an XCMSnExp object. This is in contrast to e.g. the CentWaveParam object that is passed to the actual processing method.

Usage

GenericParam(fun = character(), args = list())

## S4 method for signature 'GenericParam'
show(object)

Arguments

fun
character representing the name of the function.

args
list (ideally named) with the arguments to the function.

object
GenericParam object.

Value

The GenericParam function returns a GenericParam object.

Slots

fun

character specifying the function name.

args

list (ideally named) with the arguments to the function.

.__classVersion__

the version of the class.

Author(s)

Johannes Rainer

See Also

processHistory for how to access the process history of an XCMSnExp object.

Examples

prm <- GenericParam(fun = "mean")

prm <- GenericParam(fun = "mean", args = list(na.rm = TRUE))

[Package xcms version 3.6.1 Index]