bin,XCMSnExp-method {xcms}	R Documentation

XCMSnExp data manipulation methods inherited from MSnbase

Description

The methods listed on this page are XCMSnExp methods inherited from its parent, the OnDiskMSnExp class from the MSnbase package, that alter the raw data or are related to data subsetting. Thus calling any of these methods causes all xcms pre-processing results to be removed from the XCMSnExp object to ensure its data integrity.

bin: allows to bin spectra. See bin documentation in the MSnbase package for more details and examples.

clean: removes unused 0 intensity data points. See clean documentation in the MSnbase package for details and examples.

filterAcquisitionNum: filters the XCMSnExp object keeping only spectra with the provided acquisition numbers. See filterAcquisitionNum for details and examples.

The normalize method performs basic normalization of spectra intensities. See normalize documentation in the MSnbase package for details and examples.

The pickPeaks method performs peak picking. See pickPeaks documentation for details and examples.

The removePeaks method removes mass peaks (intensities) lower than a threshold. Note that these peaks refer to mass peaks, which are different to the chromatographic peaks detected and analyzed in a metabolomics experiment! See removePeaks documentation for details and examples.

The smooth method smooths spectra. See smooth documentation in MSnbase for details and examples.

Usage

## S4 method for signature 'XCMSnExp'
bin(object, binSize = 1L, msLevel.)

## S4 method for signature 'XCMSnExp'
clean(object, all = FALSE, verbose = FALSE,
  msLevel.)

## S4 method for signature 'XCMSnExp'
filterAcquisitionNum(object, n, file)

## S4 method for signature 'XCMSnExp'
normalize(object, method = c("max", "sum"), ...)

## S4 method for signature 'XCMSnExp'
pickPeaks(object, halfWindowSize = 3L,
  method = c("MAD", "SuperSmoother"), SNR = 0L, ...)

## S4 method for signature 'XCMSnExp'
removePeaks(object, t = "min", verbose = FALSE,
  msLevel.)

## S4 method for signature 'XCMSnExp'
smooth(x, method = c("SavitzkyGolay",
  "MovingAverage"), halfWindowSize = 2L, verbose = FALSE, ...)

Arguments

object
XCMSnExp or OnDiskMSnExp object.

binSize
numeric(1) defining the size of a bin (in Dalton).

msLevel.
For bin, clean, filterMsLevel, removePeaks: numeric(1) defining the MS level(s) to which operations should be applied or to which the object should be subsetted.

all
For clean: logical(1), if TRUE all zeros are removed.

verbose
logical(1) whether progress information should be displayed.

n
For filterAcquisitionNum: integer defining the acquisition numbers of the spectra to which the data set should be sub-setted.

file
For filterAcquisitionNum: integer defining the file index within the object to subset the object by file.

method
For normalize: character(1) specifying the normalization method. See normalize in the MSnbase package for details. For pickPeaks: character(1) defining the method. See pickPeaks for options. For smooth: character(1) defining the method. See smooth in the MSnbase package for options and details.

...
Optional additional arguments.

halfWindowSize
For pickPeaks and smooth: integer(1) defining the window size for the peak picking. See pickPeaks and smooth in the MSnbase package for details and options.

SNR
For pickPeaks: numeric(1) defining the signal to noise ratio to be considered. See pickPeaks documentation for details.

t
For removePeaks: either a numeric(1) or "min" defining the threshold (method) to be used. See removePeaks for details.

x
XCMSnExp or OnDiskMSnExp object.

Value

For all methods: a XCMSnExp object.

Author(s)

Johannes Rainer

See Also

XCMSnExp-filter for methods to filter and subset XCMSnExp objects. XCMSnExp for base class documentation. OnDiskMSnExp for the documentation of the parent class.

[Package xcms version 3.8.1 Index]