featureValues,XCMSnExp-method {xcms}	R Documentation

Accessing mz-rt feature data values

Description

featureValues,XCMSnExp : extract a matrix for feature values with rows representing features and columns samples. Parameter value allows to define which column from the chromPeaks matrix should be returned. Multiple chromatographic peaks from the same sample can be assigned to a feature. Parameter method allows to specify the method to be used in such cases to chose from which of the peaks the value should be returned.

Usage

## S4 method for signature 'XCMSnExp'
featureValues(object, method = c("medret", "maxint",
  "sum"), value = "into", intensity = "into", filled = TRUE,
  missing = NA)

Arguments

object
A XCMSnExp object providing the feature definitions.

method
character specifying the method to resolve multi-peak mappings within the same sample, i.e. to define the representative peak for a feature in samples where more than one peak was assigned to the feature. If "medret": select the peak closest to the median retention time of the feature. If "maxint": select the peak yielding the largest signal. If "sum": sum the values (only if value is "into" or "maxo".

value
character specifying the name of the column in chromPeaks(object) that should be returned. Defaults to "into" in which case the integrated peak area is returned. To get the index of the peak in the chromPeaks(object) matrix use "index".

intensity
character specifying the name of the column in the chromPeaks(objects) matrix containing the intensity value of the peak that should be used for the conflict resolution if method = "maxint".

filled
logical(1) specifying whether values for filled-in peaks should be returned or not. If filled = FALSE, an NA is returned in the matrix for the respective peak. See fillChromPeaks for details on peak filling.

missing
how missing values should be reported. Allowed values are NA (the default), a numeric or missing = "rowmin_half". The latter replaces any NA with half of the row's minimal (non-missing) value.

Value

For featureValues: a matrix with feature values, columns representing samples, rows features. The order of the features matches the order found in the featureDefinitions(object) DataFrame. The rownames of the matrix are the same than those of the featureDefinitions DataFrame. NA is reported for features without corresponding chromatographic peak in the respective sample(s).

Note

This method is equivalent to the groupval for xcmsSet objects. Note that missing = 0 should be used to get the same behaviour as groupval, i.e. report missing values as 0 after a call to fillPeaks.

Author(s)

Johannes Rainer

See Also

XCMSnExp for information on the data object.

featureDefinitions to extract the DataFrame with the feature definitions.

featureChromatograms to extract ion chromatograms for each feature.

hasFeatures to evaluate whether the XCMSnExp provides feature definitions.

groupval for the equivalent method on xcmsSet objects.

[Package xcms version 3.8.2 Index]