adjustRtime {xcms}	R Documentation

Alignment: Retention time correction methods.

Description

The adjustRtime method(s) perform retention time correction (alignment) between chromatograms of different samples. These methods are part of the modernized xcms user interface.

The implemented retention time adjustment methods are:

peakGroups

retention time correction based on aligment of features (peak groups) present in most/all samples. See adjustRtime-peakGroups for more details.

obiwarp

alignment based on the complete mz-rt data. This method does not require any identified peaks or defined features. See adjustRtime-obiwarp for more details.

Author(s)

Johannes Rainer

See Also

retcor for the old retention time correction methods. plotAdjustedRtime for visualization of alignment results.

Other retention time correction methods: adjustRtime-obiwarp, adjustRtime-peakGroups

[Package xcms version 3.8.2 Index]