calibrate-methods {xcms}	R Documentation

Calibrate peaks for correcting unprecise m/z values

Description

Calibrate peaks of a xcmsSet via a set of known masses

Arguments

object
a xcmsSet object with uncalibrated mz

calibrants
a vector or a list of vectors with reference m/z-values

method
the used calibrating-method, see below

mzppm
the relative error used for matching peaks in ppm (parts per million)

mzabs
the absolute error used for matching peaks in Da

neighbours
the number of neighbours from wich the one with the highest intensity is used (instead of the nearest)

plotres
can be set to TRUE if wanted a result-plot showing the found m/z with the distances and the regression

Value

object
a xcmsSet with one ore more samples

calibrants
for each sample different calibrants can be used, if a list of m/z-vectors is given. The length of the list must be the same as the number of samples, alternatively a single vector of masses can be given which is used for all samples.

method
"shift" for shifting each m/z, "linear" does a linear regression and adds a linear term to each m/z. "edgeshift" does a linear regression within the range of the mz-calibrants and a shift outside.

Methods

object = "xcmsSet"

calibrate(object, calibrants,method="linear", mzabs=0.0001, mzppm=5, neighbours=3, plotres=FALSE)

See Also

xcmsSet-class,

[Package xcms version 3.8.1 Index]