Extract spectra associated with chromatographic peaks

Extract (MS1 or MS2) spectra from an XCMSnExp object for each identified chromatographic peak. The function returns by default spectra for chromatographic peaks of all MS levels, but parameter peaks allows to restrict the result to selected chromatographic peaks. For msLevel = 1L (only supported for return.type = "Spectra" or return.type = "List") MS1 spectra within the retention time boundaries (in the file in which the peak was detected) are returned. For msLevel = 2L MS2 spectra are returned for a chromatographic peak if their precursor m/z is within the retention time and m/z range of the chromatographic peak. Parameter method allows to define whether all or a single spectrum should be returned:

method = "all": (default): return all spectra for each peak.

method = "closest_rt": return the spectrum with the retention time closest to the peak's retention time (at apex).

method = "closest_mz": return the spectrum with the precursor m/z closest to the peaks's m/z (at apex); only supported for msLevel = 2L.

method = "signal": return the spectrum with the sum of intensities most similar to the peak's apex signal ("maxo"); only supported for msLevel = 2L.

method = "largest_tic": return the spectrum with the largest total signal (sum of peaks intensities).

method = "largest_bpi": return the spectrum with the largest peak intensity (maximal peak intensity).

Parameter return.type allows to specify the type of the result object. Please use return.type = "Spectra" or return.type = "List", return.type = "list" or the default return.type = "MSpectra" will be deprecated (also, they do not support extracting MS1 spectra).

See also the LC-MS/MS data analysis vignette for more details and examples.

chromPeakSpectra(
  x,
  msLevel = 2L,
  expandRt = 0,
  expandMz = 0,
  ppm = 0,
  method = c("all", "closest_rt", "closest_mz", "signal", "largest_tic", "largest_bpi"),
  skipFilled = FALSE,
  return.type = c("MSpectra", "Spectra", "list", "List"),
  peaks = character()
)

Arguments

x

XCMSnExp object with identified chromatographic peaks.

msLevel

integer(1) defining whether MS1 or MS2 spectra should be returned. msLevel = 1 is currently only supported for return.type being "Spectra" or "List".

expandRt

numeric(1) to expand the retention time range of each peak by a constant value on each side.

expandMz

numeric(1) to expand the m/z range of each peak by a constant value on each side.

ppm

numeric(1) to expand the m/z range of each peak (on each side) by a value dependent on the peak's m/z.

method

character(1) specifying which spectra to include in the result. Defaults to method = "all". See function description for details.

skipFilled

logical(1) whether spectra for filled-in peaks should be reported or not.

return.type

character(1) defining the result type. Defaults to return.type = "MSpectra" but return.type = "Spectra" or return.type = "List" are preferred. See below for more information.

peaks

character, logical or integer allowing to specify a subset of chromatographic peaks in chromPeaks for which spectra should be returned (providing either their ID, a logical vector same length than nrow(chromPeaks(x)) or their index in chromPeaks(x)). This parameter overrides skipFilled and is only supported for return.type being either "Spectra" or "List".

Value

parameter return.type allow to specify the type of the returned object:

return.type = "MSpectra": a MSpectra object with elements being Spectrum objects. The result objects contains all spectra for all peaks. Metadata column "peak_id" provides the ID of the respective peak (i.e. its rowname in chromPeaks()).

return.type = "Spectra": a Spectra object (defined in the Spectra package). The result contains all spectra for all peaks. Metadata column "peak_id" provides the ID of the respective peak (i.e. its rowname in chromPeaks() and "peak_index" its index in the object's chromPeaks matrix.

return.type = "list": list of lists that are either of length 0 or contain Spectrum2 object(s) within the m/z-rt range. The length of the list matches the number of peaks.

return.type = "List": List of length equal to the number of chromatographic peaks is returned with elements being either NULL (no spectrum found) or a Spectra object.

Author

Johannes Rainer

Examples

## Read a file with DDA LC-MS/MS data
fl <- system.file("TripleTOF-SWATH/PestMix1_DDA.mzML", package = "msdata")
dda <- readMSData(fl, mode = "onDisk")

## Subset the object to reduce runtime of the example
dda <- filterRt(dda, c(200, 400))

## Perform MS1 peak detection
dda <- findChromPeaks(dda, CentWaveParam(peakwidth = c(5, 15), prefilter = c(5, 1000)))
#> Detecting mass traces at 25 ppm ... 
#> OK
#> Detecting chromatographic peaks in 7 regions of interest ...
#>  OK: 8 found.

## Load the required Spectra package and return all MS2 spectro for each
## chromatographic peaks as a Spectra object
ms2_sps <- chromPeakSpectra(dda, return.type = "Spectra")
ms2_sps
#> MSn data (Spectra) with 25 spectra in a MsBackendMzR backend:
#>            msLevel     rtime scanIndex
#>          <integer> <numeric> <integer>
#> F1.S1812         2   237.869      1812
#> F1.S1842         2   240.759      1842
#> F1.S1846         2   241.299      1846
#> F1.S2391         2   318.444      2391
#> F1.S2729         2   354.271      2729
#> ...            ...       ...       ...
#> F1.S3119         2   386.669      3119
#> F1.S3133         2   387.569      3133
#> F1.S3146         2   388.269      3146
#> F1.S2591         2   342.552      2591
#> F1.S2601         2   343.322      2601
#>  ... 38 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML

## columns peak_id or peak_index assign spectra to the chromatographic peaks
ms2_sps$peak_id
#>  [1] "CP1" "CP1" "CP1" "CP2" "CP3" "CP3" "CP3" "CP3" "CP4" "CP4" "CP4" "CP6"
#> [13] "CP6" "CP6" "CP6" "CP6" "CP6" "CP7" "CP7" "CP7" "CP7" "CP7" "CP7" "CP8"
#> [25] "CP8"
ms2_sps$peak_index
#>  [1] 1 1 1 2 3 3 3 3 4 4 4 6 6 6 6 6 6 7 7 7 7 7 7 8 8
chromPeaks(dda)
#>           mz    mzmin    mzmax      rt   rtmin   rtmax        into        intb
#> CP1 219.0961 219.0948 219.0996 240.897 235.892 246.984 15914.63661 15903.28137
#> CP2 219.0947 219.0936 219.0963 319.236 313.164 324.145    78.36274    67.09277
#> CP3 227.9913 227.9908 227.9921 357.929 351.649 365.748  5698.72930  5684.55178
#> CP4 300.0312 300.0296 300.0321 366.158 360.029 373.638 12468.01035 12454.28834
#> CP5 302.0284 302.0273 302.0295 366.158 360.449 373.638  7825.51188  7812.11153
#> CP6 224.0836 224.0831 224.0843 388.587 380.747 395.966  9401.70901  9386.10954
#> CP7 256.1102 256.1098 256.1109 388.167 380.330 395.556  8176.23790  8160.63125
#> CP8 298.2750 298.2739 298.2765 345.590 339.830 353.309 38318.95237 38305.30263
#>           maxo   sn sample
#> CP1 4877.11621 4876      1
#> CP2   19.35541   18      1
#> CP3 1459.38428 1458      1
#> CP4 3156.89258 3156      1
#> CP5 1974.66846 1974      1
#> CP6 2268.62915 2268      1
#> CP7 1836.12695 1835      1
#> CP8 8060.13037 8059      1

## Alternatively, return the result as a List of Spectra objects. This list
## is parallel to chromPeaks hence the mapping between chromatographic peaks
## and MS2 spectra is easier.
ms2_sps <- chromPeakSpectra(dda, return.type = "List")
ms2_sps[[1L]]
#> MSn data (Spectra) with 3 spectra in a MsBackendMzR backend:
#>            msLevel     rtime scanIndex
#>          <integer> <numeric> <integer>
#> F1.S1812         2   237.869      1812
#> F1.S1842         2   240.759      1842
#> F1.S1846         2   241.299      1846
#>  ... 38 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
length(ms2_sps)
#> [1] 8

## In addition to MS2 spectra we could also return the MS1 spectrum for each
## chromatographic peak which is closest to the peak's apex position.
ms1_sps <- chromPeakSpectra(dda, msLevel = 1L, method = "closest_rt",
    return.type = "Spectra")
ms1_sps
#> MSn data (Spectra) with 8 spectra in a MsBackendMzR backend:
#>            msLevel     rtime scanIndex
#>          <integer> <numeric> <integer>
#> F1.S1843         1   240.897      1843
#> F1.S2396         1   319.236      2396
#> F1.S2804         1   357.929      2804
#> F1.S2920         1   366.158      2920
#> F1.S2920         1   366.158      2920
#> F1.S3153         1   388.587      3153
#> F1.S3144         1   388.167      3144
#> F1.S2620         1   345.590      2620
#>  ... 38 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML

## Parameter peaks would allow to extract spectra for specific peaks only
chromPeakSpectra(dda, msLevel = 1L, method = "closest_rt", peaks = c(3, 5))
#> Warning: Ignoring parameter 'peaks' which is only supported for 'return.type = "Spectra"' and 'return.type = "List"'.
#> Warning: msLevel = 1 is currently not supported forreturn.type = "MSpectra" or return.type = "list".
#> MSpectra with 0 spectra and 1 metadata column(s):