descendZero {xcms}	R Documentation

Find start and end points of a peak

Description

Decends down the sides of a data peak and finds either the points greater than or equal to the zero intercept, the intercept with a given value, or the bottom of the first valley on each side.

Usage

descendZero(y, istart = which.max(y))
descendValue(y, value, istart = which.max(y))
descendMin(y, istart = which.max(y))

Arguments

y
numeric vector with values

istart
starting point for descent

value
numeric value to descend to

Value

An integer vector of length 2 with the starting and ending indicies of the peak start and end points.

Author(s)

Colin A. Smith, csmith@scripps.edu

See Also

descendValue

Examples

normdist <- dnorm(seq(-4, 4, .1)) - .1
xcms:::descendZero(normdist)
normdist[xcms:::descendZero(normdist)]
xcms:::descendValue(normdist, .15)
normdist[xcms:::descendValue(normdist, .15)]
xcms:::descendMin(normdist)

[Package xcms version 3.6.1 Index]