DEPRECATED: Extract a data.frame containing MS data

UPDATE: the extractMsData and plotMsData functions are deprecated and as(x, "data.frame") and plot(x, type = "XIC") (x being an OnDiskMSnExp or XCMSnExp object) should be used instead. See examples below. Be aware that filtering the raw object might however drop the adjusted retention times. In such cases it is advisable to use the applyAdjustedRtime() function prior to filtering.

Extract a data.frame of retention time, mz and intensity values from each file/sample in the provided rt-mz range (or for the full data range if rt and mz are not defined).

# S4 method for OnDiskMSnExp
extractMsData(object, rt, mz, msLevel = 1L)

# S4 method for XCMSnExp
extractMsData(
  object,
  rt,
  mz,
  msLevel = 1L,
  adjustedRtime = hasAdjustedRtime(object)
)

Arguments

object
A XCMSnExp or OnDiskMSnExp object.

rt
numeric(2) with the retention time range from which the data should be extracted.

mz
numeric(2) with the mz range.

msLevel
integer defining the MS level(s) to which the data should be sub-setted prior to extraction; defaults to msLevel = 1L.

adjustedRtime
(for extractMsData,XCMSnExp): logical(1) specifying if adjusted or raw retention times should be reported. Defaults to adjusted retention times, if these are present in object.

Value

A list of length equal to the number of samples/files in object. Each element being a data.frame with columns "rt", "mz" and "i" with the retention time, mz and intensity tuples of a file. If no data is available for the mz-rt range in a file a data.frame with 0 rows is returned for that file.

See also

Examples

## Read some files from the test data package.
library(faahKO)
library(xcms)
library(magrittr)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)
raw_data <- readMSData(fls[1:2], mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

## Extract the full data as a data.frame
ms_all <- as(raw_data, "data.frame")
head(ms_all)

#>   file       rt    mz    i
#> 1    1 2501.378 200.1 1716
#> 2    1 2501.378 201.0 1723
#> 3    1 2501.378 201.9 2814
#> 4    1 2501.378 202.9 1961
#> 5    1 2501.378 203.8  667
#> 6    1 2501.378 204.2  676

nrow(ms_all)

#> [1] 886563

## Read the full MS data for a defined mz-rt region.
res <- raw_data %>%
    filterRt(rt = c(2700, 2900)) %>%
    filterMz(mz = c(300, 320)) %>%
    as("data.frame")

head(res)

#>   file       rt    mz    i
#> 1    1 2700.127 300.2 1330
#> 2    1 2700.127 301.0 1221
#> 3    1 2700.127 302.1  967
#> 4    1 2700.127 303.1 3757
#> 5    1 2700.127 304.1 3482
#> 6    1 2700.127 305.1 2265

nrow(res)

#> [1] 5089