extractMsData,OnDiskMSnExp-method {xcms}	R Documentation

Extract a data.frame containing MS data

Description

Extract a data.frame of retention time, mz and intensity values from each file/sample in the provided rt-mz range (or for the full data range if rt and mz are not defined).

Usage

## S4 method for signature 'OnDiskMSnExp'
extractMsData(object, rt, mz, msLevel = 1L)

## S4 method for signature 'XCMSnExp'
extractMsData(object, rt, mz, msLevel = 1L,
  adjustedRtime = hasAdjustedRtime(object))

Arguments

object
A XCMSnExp or OnDiskMSnExp object.

rt
numeric(2) with the retention time range from which the data should be extracted.

mz
numeric(2) with the mz range.

msLevel
integer defining the MS level(s) to which the data should be sub-setted prior to extraction; defaults to msLevel = 1L.

adjustedRtime
(for extractMsData,XCMSnExp): logical(1) specifying if adjusted or raw retention times should be reported. Defaults to adjusted retention times, if these are present in object.

Value

A list of length equal to the number of samples/files in object. Each element being a data.frame with columns "rt", "mz" and "i" with the retention time, mz and intensity tuples of a file. If no data is available for the mz-rt range in a file a data.frame with 0 rows is returned for that file.

Author(s)

Johannes Rainer

See Also

XCMSnExp for the data object. plotMsData to plot the data for a single file.

Examples

## Read some files from the test data package.
library(faahKO)
library(xcms)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)
raw_data <- readMSData(fls[1:2], mode = "onDisk")

## Read the full MS data for a defined mz-rt region.
res <- extractMsData(raw_data, mz = c(300, 320), rt = c(2700, 2900))

## We've got one data.frame per file
length(res)

## With number of rows:
nrow(res[[1]])

head(res[[1]])

[Package xcms version 3.0.2 Index]