featureChromatograms {xcms}	R Documentation

Extract ion chromatograms for each feature

Description

Extract ion chromatograms for features in an XCMSnExp object. The function returns for each feature its extracted ion chromatogram and all associated peaks with it.

By default only chromatographic peaks associated with a feature are included for an extracted ion chromatogram. Setting include = "all" (instead of the default include = "feature_only") will return all chromatographic peaks identified in the m/z - rt data slice of a feature (and eventually also other features within that region).

Usage

featureChromatograms(x, expandRt = 0, aggregationFun = "max", features,
  include = c("feature_only", "all"), filled = FALSE, ...)

Arguments

x
XCMSnExp object with grouped chromatographic peaks.

expandRt
numeric(1) to expand the retention time range for each chromatographic peak by a constant value on each side.

aggregationFun
character(1) specifying the name that should be used to aggregate intensity values across the m/z value range for the same retention time. The default "sum" returns a base peak chromatogram.

features
integer, character or logical defining a subset of features for which chromatograms should be returned. Can be the index of the features in featureDefinitions, feature IDs (row names of featureDefinitions) or a logical vector.

include
character(1) defining which chromatographic peaks and feature definitions should be included in the returned XChromatograms(). See description above for details.

filled
logical(1) whether filled-in peaks should be included in the result object. The default is filled = FALSE, i.e. only detected peaks are reported.

...
optional arguments to be passed along to the chromatogram() function.

Value

XChromatograms() object.

Author(s)

Johannes Rainer

Examples

library(xcms)
library(faahKO)
faahko_3_files <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
                    system.file('cdf/KO/ko16.CDF', package = "faahKO"),
                    system.file('cdf/KO/ko18.CDF', package = "faahKO"))

## Do a simple and fast preprocessing of the test data
od <- readMSData(faahko_3_files, mode = "onDisk")
od <- findChromPeaks(od, param = CentWaveParam(peakwidth = c(30, 80),
    noise = 1000))
od <- adjustRtime(od, param = ObiwarpParam(binSize = 0.6))
od <- groupChromPeaks(od,
    param = PeakDensityParam(minFraction = 0.8, sampleGroups = rep(1, 3)))

## Extract ion chromatograms for each feature
chrs <- featureChromatograms(od)

## Plot the XIC for the first feature using different colors for each file
par(mfrow = c(1, 2))
plot(chrs[1, ], col = c("red", "green", "blue"))

[Package xcms version 3.6.1 Index]