findPeaks.MSW,xcmsRaw-method {xcms}	R Documentation

Peak detection for single-spectrum non-chromatography MS data

Description

This method performs peak detection in mass spectrometry direct injection spectrum using a wavelet based algorithm.

Usage

## S4 method for signature 'xcmsRaw'
findPeaks.MSW(object, snthresh = 3,
  verbose.columns = FALSE, ...)

Arguments

object
The xcmsRaw object on which peak detection should be performed.

snthresh
numeric(1) defining the signal to noise ratio cutoff.

verbose.columns
Logical whether additional peak meta data columns should be returned.

...
Additional parameters to be passed to the identifyMajorPeaks and sav.gol functions from the MassSpecWavelet package.

Details

This is a wrapper around the peak picker in Bioconductor's MassSpecWavelet package calling peakDetectionCWT and tuneInPeakInfo functions.

Value

A matrix, each row representing an intentified peak, with columns:

mz

m/z value of the peak at the centroid position.

mzmin

Minimum m/z of the peak.

mzmax

Maximum m/z of the peak.

rt

Always -1.

rtmin

Always -1.

rtmax

Always -1.

into

Integrated (original) intensity of the peak.

maxo

Maximum intensity of the peak.

intf

Always NA.

maxf

Maximum MSW-filter response of the peak.

sn

Signal to noise ratio.

Author(s)

Joachim Kutzera, Steffen Neumann, Johannes Rainer

See Also

MSW for the new user interface, do_findPeaks_MSW for the downstream analysis function or peakDetectionCWT from the MassSpecWavelet for details on the algorithm and additionally supported parameters.

[Package xcms version 3.8.2 Index]