Correspondence: Chromatographic peak grouping methods.

The groupChromPeaks method(s) perform the correspondence, i.e. the grouping of chromatographic peaks within and between samples. These methods are part of the modernized xcms user interface. The resulting peak groups are referred to as (mz-rt) features and can be accessed via the featureDefinitions method on the result object.

The implemented peak grouping methods are:

density

peak grouping based on time dimension peak densities. See groupChromPeaks-density for more details.

mzClust

high resolution peak grouping for single spectra (direct infusion) MS data. See groupChromPeaks-mzClust for more details.

nearest

chromatographic peak grouping based on their proximity in the mz-rt space. See groupChromPeaks-nearest for more details.

See also

Author

Johannes Rainer