groupChromPeaks {xcms}	R Documentation

Correspondence: Chromatographic peak grouping methods.

Description

The groupChromPeaks method(s) perform the correspondence, i.e. the grouping of chromatographic peaks within and between samples. These methods are part of the modernized xcms user interface. The resulting peak groups are referred to as (mz-rt) features and can be accessed via the featureDefinitions method on the result object.

The implemented peak grouping methods are:

density

peak grouping based on time dimension peak densities. See groupChromPeaks-density for more details.

mzClust

high resolution peak grouping for single spectra (direct infusion) MS data. See groupChromPeaks-mzClust for more details.

nearest

chromatographic peak grouping based on their proximity in the mz-rt space. See groupChromPeaks-nearest for more details.

Author(s)

Johannes Rainer

See Also

featureDefinitions and featureValues,XCMSnExp-method for methods to access peak grouping results.

featureChromatograms to extract ion chromatograms for each feature.

group for the old peak grouping methods.

Other peak grouping methods: groupChromPeaks-density, groupChromPeaks-mzClust, groupChromPeaks-nearest

[Package xcms version 3.8.2 Index]