Peak grouping based on time dimension peak densities

This method performs performs correspondence (chromatographic peak grouping) based on the density (distribution) of identified peaks along the retention time axis within slices of overlapping mz ranges. All peaks (from the same or from different samples) being close on the retention time axis are grouped into a feature (peak group).

The PeakDensityParam class allows to specify all settings for the peak grouping based on peak densities along the time dimension. Instances should be created with the PeakDensityParam() constructor.

sampleGroups,sampleGroups<-: getter and setter for the sampleGroups slot of the object. Its length should match the number of samples in the experiment and it should not contain NAs.

bw,bw<-: getter and setter for the bw slot of the object.

minFraction,minFraction<-: getter and setter for the minFraction slot of the object.

minSamples,minSamples<-: getter and setter for the minSamples slot of the object.

binSize,binSize<-: getter and setter for the binSize slot of the object.

maxFeatures,maxFeatures<-: getter and setter for the maxFeatures slot of the object.

groupChromPeaks,XCMSnExp,PeakDensityParam: performs correspondence (peak grouping within and across samples) within in mz dimension overlapping slices of MS data based on the density distribution of the identified chromatographic peaks in the slice along the time axis.

The correspondence analysis can be performed on chromatographic peaks of any MS level (if present and if chromatographic peak detection has been performed for that MS level) defining features combining these peaks. The MS level can be selected with the parameter msLevel. By default, calling groupChromPeaks will remove any previous correspondence results. This can be disabled with add = TRUE, which will add newly defined features to already present feature definitions.

PeakDensityParam(
  sampleGroups = numeric(),
  bw = 30,
  minFraction = 0.5,
  minSamples = 1,
  binSize = 0.25,
  maxFeatures = 50
)

# S4 method for PeakDensityParam
show(object)

# S4 method for PeakDensityParam
sampleGroups(object)

# S4 method for PeakDensityParam
sampleGroups(object) <- value

# S4 method for PeakDensityParam
bw(object)

# S4 method for PeakDensityParam
bw(object) <- value

# S4 method for PeakDensityParam
minFraction(object)

# S4 method for PeakDensityParam
minFraction(object) <- value

# S4 method for PeakDensityParam
minSamples(object)

# S4 method for PeakDensityParam
minSamples(object) <- value

# S4 method for PeakDensityParam
binSize(object)

# S4 method for PeakDensityParam
binSize(object) <- value

# S4 method for PeakDensityParam
maxFeatures(object)

# S4 method for PeakDensityParam
maxFeatures(object) <- value

# S4 method for XCMSnExp,PeakDensityParam
groupChromPeaks(object, param, msLevel = 1L, add = FALSE)

Arguments

sampleGroups
A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

bw
numeric(1) defining the bandwidth (standard deviation ot the smoothing kernel) to be used. This argument is passed to the [density() method.

minFraction
numeric(1) defining the minimum fraction of samples in at least one sample group in which the peaks have to be present to be considered as a peak group (feature).

minSamples
numeric(1) with the minimum number of samples in at least one sample group in which the peaks have to be detected to be considered a peak group (feature).

binSize
numeric(1) defining the size of the overlapping slices in mz dimension.

maxFeatures
numeric(1) with the maximum number of peak groups to be identified in a single mz slice.

object
For groupChromPeaks: an XCMSnExp object containing the results from a previous peak detection analysis (see findChromPeaks()).

For all other methods: a `PeakDensityParam` object.

value
The value for the slot.

param
A PeakDensityParam object containing all settings for the peak grouping algorithm.

msLevel
integer(1) (default msLevel = 1L) defining the MS level on which the correspondence should be performed. It is required that chromatographic peaks of the respective MS level are present.

add
logical(1) (default add = FALSE) allowing to perform an additional round of correspondence (e.g. on a different MS level) and add features to the already present feature definitions.

Value

The PeakDensityParam function returns a PeakDensityParam class instance with all of the settings specified for chromatographic peak alignment based on peak densities. Note that argument sampleGroups is mandatory and should represent either the sample grouping in the experiment. It's length has to match the number of sample in the experiments.

For groupChromPeaks: a XCMSnExp object with the results of the correspondence analysis. The definition of the resulting mz-rt features can be accessed with the featureDefinitions() method

Slots

.__classVersion__,sampleGroups,bw,minFraction,minSamples,binSize,maxFeatures

See corresponding parameter above. .__classVersion__ stores the version from the class. Slots values should exclusively be accessed via the corresponding getter and setter methods listed above.

Note

These methods and classes are part of the updated and modernized xcms user interface. All of the settings to the algorithm can be passed with a PeakDensityParam object.

References

Colin A. Smith, Elizabeth J. Want, Grace O'Maille, Ruben Abagyan and Gary Siuzdak. "XCMS: Processing Mass Spectrometry Data for Metabolite Profiling Using Nonlinear Peak Alignment, Matching, and Identification" Anal. Chem. 2006, 78:779-787.

See also

Author

Colin Smith, Johannes Rainer

Examples

## Create a PeakDensityParam object
p <- PeakDensityParam(binSize = 0.05, sampleGroups = c(1, 1, 2, 2))
## Change hte minSamples slot
minSamples(p) <- 3
p

#> Object of class:  PeakDensityParam 
#> Parameters:
#>  sampleGroups: numeric of length 4 
#>  bw: 30 
#>  minFraction: 0.5 
#>  minSamples: 3 
#>  binSize: 0.05 
#>  maxFeatures: 50 

##############################
## Chromatographic peak detection and grouping.
##
## Load a test data set with detected peaks
data(faahko_sub)
## Update the path to the files for the local system
dirname(faahko_sub) <- system.file("cdf/KO", package = "faahKO")

res <- faahko_sub

head(chromPeaks(res))

#>          mz mzmin mzmax       rt    rtmin    rtmax       into       intb   maxo
#> CP001 453.2 453.2 453.2 2506.073 2501.378 2527.982  1007409.0  1007380.8  38152
#> CP002 302.0 302.0 302.0 2617.185 2595.275 2640.659   687146.6   671297.8  30552
#> CP003 344.0 344.0 344.0 2679.783 2646.919 2709.517  5210015.9  5135916.9 152320
#> CP004 430.1 430.1 430.1 2681.348 2639.094 2712.647  2395840.3  2299899.6  65752
#> CP005 366.0 366.0 366.0 2679.783 2642.224 2718.907  3365174.0  3279468.3  79928
#> CP006 343.0 343.0 343.0 2678.218 2637.529 2712.647 24147443.2 23703761.7 672064
#>          sn sample
#> CP001 38151      1
#> CP002    46      1
#> CP003    68      1
#> CP004    42      1
#> CP005    49      1
#> CP006    87      1

## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

#> 
#>   1   2   3 
#>  87 100  61 

## Performing the chromatographic peak grouping. Assigning all samples to
## the same sample group.
fdp <- PeakDensityParam(sampleGroups = rep(1, length(fileNames(res))))
res <- groupChromPeaks(res, fdp)

#> Processing 2572 mz slices ... 

#> OK

## The definition of the features (peak groups):
featureDefinitions(res)

#> DataFrame with 47 rows and 10 columns
#>          mzmed     mzmin     mzmax     rtmed     rtmin     rtmax    npeaks
#>      <numeric> <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> FT01     279.0     279.0     279.0   2787.77   2787.76   2787.77         2
#> FT02     286.2     286.2     286.2   3254.90   3250.99   3258.82         2
#> FT03     300.2     300.2     300.2   3387.14   3379.32   3390.27         4
#> FT04     301.0     301.0     301.0   2787.77   2786.20   2792.46         3
#> FT05     305.1     305.1     305.1   2994.34   2994.34   2994.34         2
#> ...        ...       ...       ...       ...       ...       ...       ...
#> FT43    534.20     534.2     534.2   3580.41   3576.50   3584.33         2
#> FT44    536.20     536.2     536.2   3712.65   3703.26   3718.91         3
#> FT45    571.75     571.7     571.8   2918.44   2916.09   2920.78         2
#> FT46    579.10     579.1     579.1   2786.98   2786.20   2787.77         2
#> FT47    593.40     593.4     593.4   3707.18   3704.83   3709.52         2
#>             X1         peakidx  ms_level
#>      <numeric>          <list> <integer>
#> FT01         2          11,199         1
#> FT02         2         115,205         1
#> FT03         3  35,125,210,...         1
#> FT04         3      10, 97,198         1
#> FT05         2          15,203         1
#> ...        ...             ...       ...
#> FT43         2          61,227         1
#> FT44         3      73,174,237         1
#> FT45         2         100,201         1
#> FT46         2           8,197         1
#> FT47         2         164,234         1

## Using the featureValues method to extract a matrix with the
## intensities of the features per sample.
head(featureValues(res, value = "into"))

#>        ko15.CDF ko16.CDF   ko18.CDF
#> FT01 17140627.0       NA 16919266.9
#> FT02         NA  1264119   732016.2
#> FT03  4700903.2  5313736  5169558.2
#> FT04  3051847.8  1964444  2774885.3
#> FT05   847473.1       NA  1003750.6
#> FT06  1070389.9       NA  1983342.5

## The process history:
processHistory(res)

#> [[1]]
#> Object of class "XProcessHistory"
#>  type: Peak detection 
#>  date: Mon Jun 15 06:30:55 2020 
#>  info:  
#>  fileIndex: 1,2,3 
#>  Parameter class: CentWaveParam 
#>  MS level(s) 1 
#> 
#> [[2]]
#> Object of class "XProcessHistory"
#>  type: Peak grouping 
#>  date: Mon Jun 29 09:18:21 2020 
#>  info:  
#>  fileIndex: 1,2,3 
#>  Parameter class: PeakDensityParam 
#>  MS level(s) 1 
#>