Peak grouping based on proximity in the mz-rt space

This method is inspired by the grouping algorithm of mzMine (Katajamaa 2006) and performs correspondence based on proximity of peaks in the space spanned by retention time and mz values. The method creates first a master peak list consisting of all chromatographic peaks from the sample in which most peaks were identified, and starting from that, calculates distances to peaks from the sample with the next most number of peaks. If peaks are closer than the defined threshold they are grouped together.

The NearestPeaksParam class allows to specify all settings for the peak grouping based on the nearest algorithm. Instances should be created with the NearestPeaksParam constructor.

sampleGroups,sampleGroups<-: getter and setter for the sampleGroups slot of the object.

mzVsRtBalance,mzVsRtBalance<-: getter and setter for the mzVsRtBalance slot of the object.

absMz,absMz<-: getter and setter for the absMz slot of the object.

absRt,absRt<-: getter and setter for the absRt slot of the object.

kNN,kNN<-: getter and setter for the kNN slot of the object.

groupChromPeaks,XCMSnExp,NearestPeaksParam: performs peak grouping based on the proximity between chromatographic peaks from different samples in the mz-rt range.

The correspondence analysis can be performed on chromatographic peaks of any MS level (if present and if chromatographic peak detection has been performed for that MS level) defining features combining these peaks. The MS level can be selected with the parameter msLevel. By default, calling groupChromPeaks will remove any previous correspondence results. This can be disabled with add = TRUE, which will add newly defined features to already present feature definitions.

NearestPeaksParam(
  sampleGroups = numeric(),
  mzVsRtBalance = 10,
  absMz = 0.2,
  absRt = 15,
  kNN = 10
)

# S4 method for NearestPeaksParam
show(object)

# S4 method for NearestPeaksParam
sampleGroups(object)

# S4 method for NearestPeaksParam
sampleGroups(object) <- value

# S4 method for NearestPeaksParam
mzVsRtBalance(object)

# S4 method for NearestPeaksParam
mzVsRtBalance(object) <- value

# S4 method for NearestPeaksParam
absMz(object)

# S4 method for NearestPeaksParam
absMz(object) <- value

# S4 method for NearestPeaksParam
absRt(object)

# S4 method for NearestPeaksParam
absRt(object) <- value

# S4 method for NearestPeaksParam
kNN(object)

# S4 method for NearestPeaksParam
kNN(object) <- value

# S4 method for XCMSnExp,NearestPeaksParam
groupChromPeaks(object, param, msLevel = 1L, add = FALSE)

Arguments

sampleGroups
A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

mzVsRtBalance
numeric(1) representing the factor by which mz values are multiplied before calculating the (euclician) distance between two peaks.

absMz
numeric(1) maximum tolerated distance for mz values.

absRt
numeric(1) maximum tolerated distance for rt values.

kNN
numeric(1) representing the number of nearest neighbors to check.

object
For groupChromPeaks: an XCMSnExp object containing the results from a previous chromatographic peak detection analysis (see findChromPeaks()).

For all other methods: a `NearestPeaksParam` object.

value
The value for the slot.

param
A PeakDensityParam object containing all settings for the peak grouping algorithm.

msLevel
integer(1) defining the MS level. Currently only MS level 1 is supported.

add
logical(1) (default add = FALSE) allowing to perform an additional round of correspondence (e.g. on a different MS level) and add features to the already present feature definitions.

Value

The NearestPeaksParam function returns a NearestPeaksParam class instance with all of the settings specified for peak alignment based on peak proximity.

For groupChromPeaks: a XCMSnExp object with the results of the peak grouping/correspondence step (i.e. the mz-rt features). These can be accessed with the featureDefinitions() method.

Slots

.__classVersion__,sampleGroups,mzVsRtBalance,absMz,absRt,kNN

See corresponding parameter above. .__classVersion__ stores the version from the class. Slots values should exclusively be accessed via the corresponding getter and setter methods listed above.

Note

These methods and classes are part of the updated and modernized xcms user interface. All of the settings to the algorithm can be passed with a NearestPeaksParam object.

References

Katajamaa M, Miettinen J, Oresic M: MZmine: Toolbox for processing and visualization of mass spectrometry based molecular profile data. Bioinformatics 2006, 22:634-636.

See also

Examples

## Create a NearestPeaksParam object
p <- NearestPeaksParam(kNN = 3)
p

#> Object of class:  NearestPeaksParam 
#> Parameters:
#>  sampleGroups: numeric of length 0 
#>  mzVsRtBalance: 10 
#>  absMz: 0.2 
#>  absRt: 15 
#>  kNN: 3 

##############################
## Chromatographic peak detection and grouping.
##
## Below we perform first a chromatographic peak detection (using the
## matchedFilter method) on some of the test files from the faahKO package
## followed by a peaks grouping using the "nearest" method.
library(faahKO)
library(MSnbase)
fls <- dir(system.file("cdf/KO", package = "faahKO"), recursive = TRUE,
           full.names = TRUE)

## Reading 2 of the KO samples
raw_data <- readMSData(fls[1:2], mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

## Perform the peak detection using the matchedFilter method.
mfp <- MatchedFilterParam(snthresh = 20, binSize = 1)
res <- findChromPeaks(raw_data, param = mfp)

head(chromPeaks(res))

#>             mz mzmin mzmax       rt    rtmin    rtmax      into      intf  maxo
#> CP001 205.0000   205 205.0 2784.635 2770.550 2800.284 1778568.9 3589494.1 84280
#> CP002 205.0000   205 205.0 2784.635 2770.550 2800.284 1778568.9 3589494.1 84280
#> CP003 207.0821   207 207.1 2712.647 2698.562 2726.731  380873.0  725545.6 18800
#> CP004 207.0821   207 207.1 2712.647 2698.562 2726.731  380873.0  725657.0 18800
#> CP005 233.0390   233 233.1 3019.378 3005.293 3033.462  399145.3  746572.1 19752
#> CP006 233.0390   233 233.1 3019.378 3005.293 3033.462  399145.3  746587.0 19752
#>            maxf i       sn sample
#> CP001 194326.10 1 88.11283      1
#> CP002 194326.10 1 87.30418      1
#> CP003  39851.35 1 25.05928      1
#> CP004  39851.56 1 26.10968      1
#> CP005  41455.94 1 28.58299      1
#> CP006  41456.03 1 30.40953      1

## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

#> 
#>   1   2 
#> 251 299 

## Performing the peak grouping
p <- NearestPeaksParam()
res <- groupChromPeaks(res, param = p)

#> Empty 'sampleGroups' in 'param', assuming all samples to be in the same group.

#> Processing sample number 1 ... 

#> OK

## The results from the peak grouping:
featureDefinitions(res)

#> DataFrame with 471 rows and 10 columns
#>           mzmed     mzmin     mzmax     rtmed     rtmin     rtmax    npeaks
#>       <numeric> <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> FT001   200.100   200.100   200.100   2876.97   2876.97   2876.97         1
#> FT002   201.091   201.091   201.091   3141.44   3141.44   3141.44         1
#> FT003   201.091   201.091   201.091   3141.44   3141.44   3141.44         1
#> FT004   205.000   205.000   205.000   2787.76   2784.64   2790.89         2
#> FT005   205.000   205.000   205.000   2790.89   2790.89   2790.89         1
#> ...         ...       ...       ...       ...       ...       ...       ...
#> FT467   594.313   594.313   594.313   3603.11   3603.11   3603.11         1
#> FT468   594.313   594.313   594.313   3603.11   3603.11   3603.11         1
#> FT469   596.317   596.317   596.317   3797.16   3797.16   3797.16         1
#> FT470   599.313   599.313   599.313   3743.95   3743.95   3743.95         1
#> FT471   599.313   599.313   599.313   3743.95   3743.95   3743.95         1
#>              X1 peakidx  ms_level
#>       <numeric>  <list> <integer>
#> FT001         1     252         1
#> FT002         1     253         1
#> FT003         1     254         1
#> FT004         2   1,255         1
#> FT005         1     256         1
#> ...         ...     ...       ...
#> FT467         1     247         1
#> FT468         1     248         1
#> FT469         1     249         1
#> FT470         1     250         1
#> FT471         1     251         1

## Using the featureValues method to extract a matrix with the intensities of
## the features per sample.
head(featureValues(res, value = "into"))

#>       ko15.CDF  ko16.CDF
#> FT001       NA  451600.7
#> FT002       NA  272872.0
#> FT003       NA  272872.0
#> FT004  1778569 1567038.1
#> FT005       NA 1567038.1
#> FT006   380873  449392.6

## The process history:
processHistory(res)

#> [[1]]
#> Object of class "XProcessHistory"
#>  type: Peak detection 
#>  date: Mon Jun 15 14:29:27 2020 
#>  info:  
#>  fileIndex: 1,2 
#>  Parameter class: MatchedFilterParam 
#>  MS level(s) 1 
#> 
#> [[2]]
#> Object of class "XProcessHistory"
#>  type: Peak grouping 
#>  date: Mon Jun 15 14:29:29 2020 
#>  info:  
#>  fileIndex: 1,2 
#>  Parameter class: NearestPeaksParam 
#>  MS level(s) 1 
#>