Peak grouping based on proximity in the mz-rt space

This method is inspired by the grouping algorithm of mzMine (Katajamaa 2006) and performs correspondence based on proximity of peaks in the space spanned by retention time and mz values. The method creates first a master peak list consisting of all chromatographic peaks from the sample in which most peaks were identified, and starting from that, calculates distances to peaks from the sample with the next most number of peaks. If peaks are closer than the defined threshold they are grouped together.

The NearestPeaksParam class allows to specify all settings for the peak grouping based on the nearest algorithm. Instances should be created with the NearestPeaksParam constructor.

sampleGroups,sampleGroups<-: getter and setter for the sampleGroups slot of the object.

mzVsRtBalance,mzVsRtBalance<-: getter and setter for the mzVsRtBalance slot of the object.

absMz,absMz<-: getter and setter for the absMz slot of the object.

absRt,absRt<-: getter and setter for the absRt slot of the object.

kNN,kNN<-: getter and setter for the kNN slot of the object.

groupChromPeaks,XCMSnExp,NearestPeaksParam: performs peak grouping based on the proximity between chromatographic peaks from different samples in the mz-rt range.

The correspondence analysis can be performed on chromatographic peaks of any MS level (if present and if chromatographic peak detection has been performed for that MS level) defining features combining these peaks. The MS level can be selected with the parameter msLevel. By default, calling groupChromPeaks will remove any previous correspondence results. This can be disabled with add = TRUE, which will add newly defined features to already present feature definitions.

NearestPeaksParam(
  sampleGroups = numeric(),
  mzVsRtBalance = 10,
  absMz = 0.2,
  absRt = 15,
  kNN = 10
)

# S4 method for NearestPeaksParam
show(object)

# S4 method for NearestPeaksParam
sampleGroups(object)

# S4 method for NearestPeaksParam
sampleGroups(object) <- value

# S4 method for NearestPeaksParam
mzVsRtBalance(object)

# S4 method for NearestPeaksParam
mzVsRtBalance(object) <- value

# S4 method for NearestPeaksParam
absMz(object)

# S4 method for NearestPeaksParam
absMz(object) <- value

# S4 method for NearestPeaksParam
absRt(object)

# S4 method for NearestPeaksParam
absRt(object) <- value

# S4 method for NearestPeaksParam
kNN(object)

# S4 method for NearestPeaksParam
kNN(object) <- value

# S4 method for XCMSnExp,NearestPeaksParam
groupChromPeaks(object, param, msLevel = 1L, add = FALSE)

Arguments

sampleGroups
A vector of the same length than samples defining the sample group assignments (i.e. which samples belong to which sample group). This parameter is mandatory for the PeakDensityParam and has to be provided also if there is no sample grouping in the experiment (in which case all samples should be assigned to the same group).

mzVsRtBalance
numeric(1) representing the factor by which mz values are multiplied before calculating the (euclician) distance between two peaks.

absMz
numeric(1) maximum tolerated distance for mz values.

absRt
numeric(1) maximum tolerated distance for rt values.

kNN
numeric(1) representing the number of nearest neighbors to check.

object
For groupChromPeaks: an XCMSnExp object containing the results from a previous chromatographic peak detection analysis (see findChromPeaks()).

For all other methods: a `NearestPeaksParam` object.

value
The value for the slot.

param
A PeakDensityParam object containing all settings for the peak grouping algorithm.

msLevel
integer(1) defining the MS level. Currently only MS level 1 is supported.

add
logical(1) (default add = FALSE) allowing to perform an additional round of correspondence (e.g. on a different MS level) and add features to the already present feature definitions.

Value

The NearestPeaksParam function returns a NearestPeaksParam class instance with all of the settings specified for peak alignment based on peak proximity.

For groupChromPeaks: a XCMSnExp object with the results of the peak grouping/correspondence step (i.e. the mz-rt features). These can be accessed with the featureDefinitions() method.

Slots

.__classVersion__,sampleGroups,mzVsRtBalance,absMz,absRt,kNN

See corresponding parameter above. .__classVersion__ stores the version from the class. Slots values should exclusively be accessed via the corresponding getter and setter methods listed above.

Note

These methods and classes are part of the updated and modernized xcms user interface. All of the settings to the algorithm can be passed with a NearestPeaksParam object.

References

Katajamaa M, Miettinen J, Oresic M: MZmine: Toolbox for processing and visualization of mass spectrometry based molecular profile data. Bioinformatics 2006, 22:634-636.

See also

Examples

## Create a NearestPeaksParam object
p <- NearestPeaksParam(kNN = 3)
p

#> Object of class:  NearestPeaksParam 
#> Parameters:
#>  sampleGroups: numeric of length 0 
#>  mzVsRtBalance: 10 
#>  absMz: 0.2 
#>  absRt: 15 
#>  kNN: 3 

## Load a test data set with detected peaks
data(faahko_sub)
## Update the path to the files for the local system
dirname(faahko_sub) <- system.file("cdf/KO", package = "faahKO")
res <- faahko_sub

head(chromPeaks(res))

#>          mz mzmin mzmax       rt    rtmin    rtmax       into       intb   maxo
#> CP001 453.2 453.2 453.2 2506.073 2501.378 2527.982  1007409.0  1007380.8  38152
#> CP002 302.0 302.0 302.0 2617.185 2595.275 2640.659   687146.6   671297.8  30552
#> CP003 344.0 344.0 344.0 2679.783 2646.919 2709.517  5210015.9  5135916.9 152320
#> CP004 430.1 430.1 430.1 2681.348 2639.094 2712.647  2395840.3  2299899.6  65752
#> CP005 366.0 366.0 366.0 2679.783 2642.224 2718.907  3365174.0  3279468.3  79928
#> CP006 343.0 343.0 343.0 2678.218 2637.529 2712.647 24147443.2 23703761.7 672064
#>          sn sample
#> CP001 38151      1
#> CP002    46      1
#> CP003    68      1
#> CP004    42      1
#> CP005    49      1
#> CP006    87      1

## The number of peaks identified per sample:
table(chromPeaks(res)[, "sample"])

#> 
#>   1   2   3 
#>  87 100  61 

## Performing the peak grouping
p <- NearestPeaksParam()
res <- groupChromPeaks(res, param = p)

#> Empty 'sampleGroups' in 'param', assuming all samples to be in the same group.

#> Processing sample number 1 ... 

#> OK

#> Processing sample number 3 ... 

#> OK

## The results from the peak grouping:
featureDefinitions(res)

#> DataFrame with 188 rows and 10 columns
#>           mzmed     mzmin     mzmax     rtmed     rtmin     rtmax    npeaks
#>       <numeric> <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> FT001     384.1     384.1     384.1   2603.10   2603.10   2603.10         1
#> FT002     533.3     533.3     533.3   2686.04   2686.04   2686.04         1
#> FT003     344.0     344.0     344.0   2682.91   2679.78   2686.04         3
#> FT004     366.0     366.0     366.0   2682.91   2679.78   2686.04         2
#> FT005     343.0     343.0     343.0   2684.48   2678.22   2686.04         3
#> ...         ...       ...       ...       ...       ...       ...       ...
#> FT184     283.2     283.2     283.2   3873.84   3873.84   3873.84         1
#> FT185     536.0     536.0     536.0   3927.05   3927.05   3927.05         1
#> FT186     411.3     411.3     411.3   3934.88   3934.88   3934.88         1
#> FT187     426.2     426.2     426.2   3794.03   3794.03   3794.03         1
#> FT188     384.2     384.2     384.2   3986.52   3986.52   3986.52         1
#>              X1     peakidx  ms_level
#>       <numeric>      <list> <integer>
#> FT001         1          88         1
#> FT002         1          89         1
#> FT003         3   3, 90,193         1
#> FT004         2        5,91         1
#> FT005         3   6, 92,195         1
#> ...         ...         ...       ...
#> FT184         1         242         1
#> FT185         1         244         1
#> FT186         1         245         1
#> FT187         1         246         1
#> FT188         1         247         1

## Using the featureValues method to extract a matrix with the intensities of
## the features per sample.
head(featureValues(res, value = "into"))

#>       ko15.CDF   ko16.CDF ko18.CDF
#> FT001       NA   274906.3       NA
#> FT002       NA    70456.3       NA
#> FT003  5210016  5700652.0  5255689
#> FT004  3365174  3448376.1       NA
#> FT005 24147443 26229546.1 23821647
#> FT006 14975761 15565868.0 12388547

## The process history:
processHistory(res)

#> [[1]]
#> Object of class "XProcessHistory"
#>  type: Peak detection 
#>  date: Mon Jun 15 06:30:55 2020 
#>  info:  
#>  fileIndex: 1,2,3 
#>  Parameter class: CentWaveParam 
#>  MS level(s) 1 
#> 
#> [[2]]
#> Object of class "XProcessHistory"
#>  type: Peak grouping 
#>  date: Mon Jun 29 09:18:24 2020 
#>  info:  
#>  fileIndex: 1,2,3 
#>  Parameter class: NearestPeaksParam 
#>  MS level(s) 1 
#>