Identify overlapping features

overlappingFeatures identifies features that are overlapping or close in the m/z - rt space.

overlappingFeatures(x, expandMz = 0, expandRt = 0, ppm = 0)

Arguments

x
XCMSnExp with the features.

expandMz
numeric(1) with the value to expand each feature (on each side) in m/z dimension before identifying overlapping features. The resulting "mzmin" for the feature is thus mzmin - expandMz and the "mzmax" mzmax + expandMz.

expandRt
numeric(1) with the value to expand each feature (on each side) in retention time dimension before identifying overlapping features. The resulting "rtmin" for the feature is thus rtmin - expandRt and the "rtmax" rtmax + expandRt.

ppm
numeric(1) to grow the m/z width of the feature by a relative value: mzmin - mzmin * ppm / 2e6, mzmax + mzmax * ppm / 2e6. Each feature is thus expanded in m/z dimension by ppm/2 on each side before identifying overlapping features.

Value

list with indices of features (in featureDefinitions()) that are overlapping.

Examples

## Load 2 test files.
data <- readMSData(c(system.file("cdf/KO/ko15.CDF", package = "faahKO"),
                     system.file("cdf/KO/ko16.CDF", package = "faahKO")),
                   mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

## Perform peak detection; parameters set to reduce processing speed
data <- findChromPeaks(data, CentWaveParam(prefilter = c(3, 10000),
    noise = 10000, snthresh = 40))

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 937 regions of interest ...

#>  OK: 87 found.

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 1025 regions of interest ...

#>  OK: 100 found.

## Correspondence analysis
data <- groupChromPeaks(data, param = PeakDensityParam(sampleGroups = c(1, 1)))

#> Processing 2572 mz slices ... 

#> OK

## Identify overlapping features
overlappingFeatures(data)

#> list()

## Identify features that are separated on retention time by less than
## 2 minutes
overlappingFeatures(data, expandRt = 60)

#> [[1]]
#> [1] 14 15
#> 
#> [[2]]
#> [1] 66 67 68 69 70
#> 
#> [[3]]
#> [1] 80 81
#>