overlappingFeatures {xcms}	R Documentation

Identify overlapping features

Description

overlappingFeatures identifies features that are overlapping or close in the m/z - rt space.

Usage

overlappingFeatures(x, expandMz = 0, expandRt = 0, ppm = 0)

Arguments

x
XCMSnExp with the features.

expandMz
numeric(1) with the value to expand each feature (on each side) in m/z dimension before identifying overlapping features. The resulting "mzmin" for the feature is thus mzmin - expandMz and the "mzmax" mzmax + expandMz.

expandRt
numeric(1) with the value to expand each feature (on each side) in retention time dimension before identifying overlapping features. The resulting "rtmin" for the feature is thus rtmin - expandRt and the "rtmax" rtmax + expandRt.

ppm
numeric(1) to grow the m/z width of the feature by a relative value: mzmin - mzmin * ppm / 2e6, mzmax + mzmax * ppm / 2e6. Each feature is thus expanded in m/z dimension by ppm/2 on each side before identifying overlapping features.

Value

list with indices of features (in featureDefinitions()) that are overlapping.

Author(s)

Johannes Rainer

Examples

## Load 2 test files.
data <- readMSData(c(system.file("cdf/KO/ko15.CDF", package = "faahKO"),
                     system.file("cdf/KO/ko16.CDF", package = "faahKO")),
                   mode = "onDisk")

## Perform peak detection; parameters set to reduce processing speed
data <- findChromPeaks(data, CentWaveParam(noise = 10000, snthresh = 40))

## Correspondence analysis
data <- groupChromPeaks(data, param = PeakDensityParam(sampleGroups = c(1, 1)))

## Identify overlapping features
overlappingFeatures(data)

## Identify features that are separated on retention time by less than
## 2 minutes
overlappingFeatures(data, expandRt = 60)

[Package xcms version 3.6.1 Index]