Plot feature groups in the m/z-retention time space

plotFeatureGroups visualizes defined feature groups in the m/z by retention time space. Features are indicated by points with features from the same feature group being connected by a line. See featureGroups() for details on and options for feature grouping.

plotFeatureGroups(
  x,
  xlim = numeric(),
  ylim = numeric(),
  xlab = "retention time",
  ylab = "m/z",
  pch = 4,
  col = "#00000060",
  type = "o",
  main = "Feature groups",
  featureGroups = character()
)

Arguments

x

XCMSnExp() object with grouped features (i.e. after calling groupFeatures().

xlim

numeric(2) with the lower and upper limit for the x-axis.

ylim

numeric(2) with the lower and upper limit for the y-axis.

xlab

character(1) with the label for the x-axis.

ylab

character(1) with the label for the y-axis.

pch

the plotting character. Defaults to pch = 4 i.e. plotting features as crosses. See par() for more information.

col

color to be used to draw the features. At present only a single color is supported.

type

plotting type (see par()). Defaults to type = "o" which draws each feature as a point and connecting the features of the same feature group with a line.

main

character(1) with the title of the plot.

featureGroups

optional character of feature group IDs to draw only specified feature group(s). If not provided, all feature groups are drawn.

Author

Johannes Rainer