plotMsData {xcms}	R Documentation

Create a plot that combines a XIC and a mz/rt 2D plot for one sample

Description

The plotMsData creates a plot that combines an (base peak ) extracted ion chromatogram on top (rt against intensity) and a plot of rt against m/z values at the bottom.

Usage

plotMsData(x, main = "", cex = 1, mfrow = c(2, 1),
  grid.color = "lightgrey", colramp = colorRampPalette(rev(brewer.pal(9,
  "YlGnBu"))))

Arguments

x
data.frame such as returned by the extractMsData() function. Only a single data.frame is supported.

main
character(1) specifying the title.

cex
numeric(1) defining the size of points. Passed directly to the plot function.

mfrow
numeric(2) defining the plot layout. This will be passed directly to par(mfrow = mfrow). See par for more information. Setting mfrow = NULL avoids calling par(mfrow = mfrow) hence allowing to pre-define the plot layout.

grid.color
a color definition for the grid line (or NA to skip creating them).

colramp
a color ramp palette to be used to color the data points based on their intensity. See argument col.regions in lattice::level.colors documentation.

Author(s)

Johannes Rainer

See Also

extractMsData() for the method to extract the data to plot.

Examples

## Read two files from the faahKO package
library(faahKO)
cdfs <- dir(system.file("cdf", package = "faahKO"), full.names = TRUE,
    recursive = TRUE)[1:2]
raw_data <- readMSData(cdfs, mode = "onDisk")
## Extract the MS data from a slice of data
msd <- extractMsData(raw_data, mz = c(334.9, 335.1), rt = c(2700, 2900))

## Plot the data for the first file
plotMsData(msd[[1]])

## To plot the data for both files:
layout(mat = matrix(1:4, ncol = 2))
plotMsData(msd[[1]], mfrow = NULL)
plotMsData(msd[[2]], mfrow = NULL)

[Package xcms version 3.0.2 Index]