xcmsEIC-class {xcms}	R Documentation

Class xcmsEIC, a class for multi-sample extracted ion chromatograms

Description

This class is used to store and plot parallel extracted ion chromatograms from multiple sample files. It integrates with the xcmsSet class to display peak area integrated during peak identification or fill-in.

Objects from the Class

Objects can be created with the getEIC method of the xcmsSet class. Objects can also be created by calls of the form new("xcmsEIC", ...).

Slots

eic:

list containing named entries for every sample. for each entry, a list of two column EIC matricies with retention time and intensity

mzrange:

two column matrix containing starting and ending m/z for each EIC

rtrange:

two column matrix containing starting and ending time for each EIC

rt:

either "raw" or "corrected" to specify retention times contained in the object

groupnames:

group names from xcmsSet object used to generate EICs

Methods

groupnames

signature(object = "xcmsEIC"): get groupnames slot

mzrange

signature(object = "xcmsEIC"): get mzrange slot

plot

signature(x = "xcmsEIC"): plot the extracted ion chromatograms

rtrange

signature(object = "xcmsEIC"): get rtrange slot

sampnames

signature(object = "xcmsEIC"): get sample names

Note

No notes yet.

Author(s)

Colin A. Smith, csmith@scripps.edu

See Also

getEIC

[Package xcms version 3.8.2 Index]